% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trees_functions.R
\name{addtrees_hadukdata}
\alias{addtrees_hadukdata}
\title{Get HadUK data from ceda archive on jasmine}
\usage{
addtrees_hadukdata(
  startdate,
  enddate,
  dtmc,
  basepath = "",
  varn = c("rainfall", "tasmax", "tasmin", "hurs", "psl", "pv", "sun", "sfsWind"),
  freq = c("day", "mon")
)
}
\arguments{
\item{startdate}{\itemize{
\item POSIXlt value indicating start date for data required
}}

\item{enddate}{\itemize{
\item POSIXlt value indicating end date for data required
}}

\item{dtmc}{\itemize{
\item spatraster coarse dtm
}}

\item{basepath}{\itemize{
\item "" for use with jasmin
}}

\item{varn}{vector of variables required, one or more of: 'rainfall','tasmax','tasmin','hurs','psl','pv','sun','sfsWind'.
NOTE: Only \code{rainfall} (precipitation in mm), \code{tasmax}(maximum daily temperature, deg C) and
\code{tasmin} (minimum daily temperature, deg C) are available at daily time steps (all other monthly only).}

\item{freq}{\itemize{
\item string of frequency required - either 'day' or 'mon' (monthly)
}}
}
\value{
database detailing download details and success of all requested files
}
\description{
Get HadUK data from ceda archive on jasmine
}
\seealso{
\code{\link[=download_hadukdaily]{download_hadukdaily()}}
\dontrun{

}
}
\keyword{jasmin}
