% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{atmos_to_sea_pressure}
\alias{atmos_to_sea_pressure}
\title{Convert atmospheric to sea level pressure}
\usage{
atmos_to_sea_pressure(pres, dtm)
}
\arguments{
\item{pres}{\itemize{
\item vector or 3Darray or spatraster of atmospheric pressure - if array expects format\link{x,y,time}
}}

\item{dtm}{\itemize{
\item vector or matrix or spatraster of elevations - if matrix or spatraster expected to match pres
}}
}
\value{
numeric or spatraster of atmospheric pressures matching dtm
}
\description{
Convert atmospheric to sea level pressure
}
\examples{
atmos_to_sea_pressure(ukcpinput$pres,ukcpinput$dtm)
}
\keyword{climate}
