% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{basindelin_big}
\alias{basindelin_big}
\title{equivalent of \code{\link[=basindelin]{basindelin()}} for large areas}
\usage{
basindelin_big(dtm, boundary = 0, tilesize = 100, plotprogress = FALSE)
}
\arguments{
\item{dtm}{a SpatRast object of elevations}

\item{boundary}{optional numeric value. If greater than 0, adjoining basins
separated by elevation differences < boundary are merged (see details.}

\item{tilesize}{optionally, the size of each tile over which  \code{\link[=basindelin]{basindelin()}}
is applied.}

\item{plotprogress}{optional logical indicating to whether to plot results for
each column as calculations progress.}
}
\value{
a SpatRast of basins sequentially numbered as integers.
}
\description{
The function \code{basindelin_big} applies \code{\link[=basindelin]{basindelin()}} in tiles and
merges the tiles
}
\examples{
bsn<-basindelin(rast(dtmf)) # Takes ~30 seconds to run
plot(bsn)
}
\seealso{
\code{\link[=basindelin]{basindelin()}}
}
