% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{biascorrect}
\alias{biascorrect}
\title{Applies bias correction to one climate variable}
\usage{
biascorrect(hist_obs, hist_mod, fut_mod, rangelims = NA)
}
\arguments{
\item{hist_obs}{a 3D array or stacked SpatRast of climate observations. If an
array the the first two dimension must represent and the third time.}

\item{hist_mod}{a 3D array or stacked SpatRast of modelled climate data for
for the same period and spatial area as \code{hist_obs} \code{dim(hist_mod)} must equal
\code{dim(hist_obs)}}

\item{fut_mod}{a 3D array or SpatRast of modelled climate data for
for e.g. a future period to which corrections are applied. Must have the same
x and y dims as \code{hist_obs} and \code{hist_mod}, but can cover a shorter or
longer time periods.}

\item{rangelims}{fractional amount above or below the range of values in
\code{hist_obs} by which values in the returned dataset can vary. I.e. if set to
1.1, values in any given grid cell of the returned dataset cannot exceed
1.1 times the range of values in the corresponding grid cell in \code{hist_obs}.
If not supplied (the default), returned values are unbounded.}
}
\value{
if \code{hist_obs} is a SpatRast, a SpatRast of bias corrected data. If
\code{hist_obs} is a 3D array, a 3D array of bias corrected data.
}
\description{
The function \code{biascorrect} applies bias correction to spatial
climate dataset
}
\seealso{
\code{\link[=precipcorrect]{precipcorrect()}} for applying corrections to precipitation data.
}
