% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect_functions.R
\name{biascorrect_apply}
\alias{biascorrect_apply}
\title{Applies bias correction model to climate variable}
\usage{
biascorrect_apply(fut_mod, biasmods, rangelims = NA)
}
\arguments{
\item{fut_mod}{a SpatRaster of modelled climate data for for e.g. a future
period to which corrections are applied.}

\item{biasmods}{an object of class \code{biascorrectmodels} derived using \code{\link[=biascorrect]{biascorrect()}}.}
}
\value{
a SpatRaster matching \code{fut_mod} with climate data bias corrected.
}
\description{
The function \code{biascorrect_apply} applies a set of bias corrections
derived using \code{bias_correct} to a spatial climate dataset
}
\details{
The \code{biascorrect_apply} function allows the same bias correction to be applied
to multiple datasets, spanning e.g. several future time periods or climate scenarious.
The spatial extent of \code{fut_mod} must at least partially overlap with that of
\code{hist_obs} when running \code{biascorrect} as corrected values are only returned for the overlapping area
}
\seealso{
\code{\link[=biascorrect]{biascorrect()}} for deriving \code{biasmods} and \code{precip_correct} for
applying corrections to precipitation datasets.
}
\keyword{biascorrect}
