% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{calculate_windcoeffs}
\alias{calculate_windcoeffs}
\title{Calculate wind coefficients - for use in wind downscaling}
\usage{
calculate_windcoeffs(dtmc, dtmm, dtmf, zi = 10, zo = 2, toArray = TRUE)
}
\arguments{
\item{dtmc}{\itemize{
\item coarse resolution raster overing wider extent than dtmf
}}

\item{dtmm}{\itemize{
\item medium resolution raster covering wider extent than dtmf
}}

\item{dtmf}{\itemize{
\item fine scale raster of downscaling area
}}

\item{zo}{\itemize{
\item wind height in metres of coeeficients
}}

\item{toArray}{\itemize{
\item if FALSE will produce a spatraster stack otherwise as 3D array
}}
}
\value{
wca2 - array of wind coeeficients for each of 8 wind directions
}
\description{
Calculate wind coefficients - for use in wind downscaling
}
