% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{checkinputs}
\alias{checkinputs}
\title{Check and summarise coarse resolution climate data inputs}
\usage{
checkinputs(input_list, tstep = c("hour", "day"), plots = TRUE)
}
\arguments{
\item{input_list}{list of climate and associated variables as output by functions like 'ukcp18toclimarray'}

\item{tstep}{string decribing whether input data is by hour or day}
}
\value{
UNMODIFIED input_list - but may print warnings and summary stats and plots of each variable if requested
}
\description{
Check and summarise coarse resolution climate data inputs
}
\examples{
climdata<- read_climdata(mesoclim::ukcpinput)
checkinputs(climdata, tstep = "day")
}
\keyword{preprocess}
