% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{clearskyrad}
\alias{clearskyrad}
\title{Calculates clear sky radiation}
\usage{
clearskyrad(tme, lat, long, tc = 15, rh = 80, pk = 101.3)
}
\arguments{
\item{tme}{a POSIXlt object of times}

\item{lat}{latitude (decimal degrees)}

\item{long}{longitude (decimal degrees)}

\item{tc}{temperature (deg C)}

\item{rh}{relative humidity (percentage)}

\item{pk}{atmospheric pressure (kPa)}
}
\value{
expected clear-sky radiation (W/m^2)
}
\description{
Calculates clear sky radiation
}
