% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{converthumidity}
\alias{converthumidity}
\title{Converts between different humidity types}
\usage{
converthumidity(
  h,
  intype = "relative",
  outtype = "vapour pressure",
  tc = 11,
  pk = 101.3
)
}
\arguments{
\item{intype}{\itemize{
\item one of relative, absolute, specific or vapour pressure
}}

\item{outtype}{\itemize{
\item one of relative, absolute, specific or vapour pressure
}}

\item{tc}{\itemize{
\item temperature
}}

\item{pk}{\itemize{
\item surface pressure
}}

\item{h-}{humidity}
}
\value{
returns humidity
(Percentage for relative,Kg / Kg for specific, kg / m3 for absolute and kPa for vapour pressure)
}
\description{
Converts between different humidity types
}
\examples{
rh<-c(25,50,75,90,100)
vp<-round(converthumidity(rh),3)
print(paste(rh,' relative humidity converts to',vp,'vapour pressure'))
}
