% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{correct_ukcp_dates}
\alias{correct_ukcp_dates}
\title{Corrects time series of UKCP18 360 day year data to valid calendar dates}
\usage{
correct_ukcp_dates(ukcp_dates)
}
\arguments{
\item{ukcp_dates}{vector of POSIXlt UKCP18 dates  as returned by 'get_ukcp18_dates()'.}
}
\value{
a vector of valid date strings in the form 'yyy-dd-mm'
}
\description{
Converts 12 x 30 day monthly data provided by UKCP18 to data corresponding to actual calendar dates
by reassigning invalid dates (eg 29-30 Feb). Does NOT add missing dates (eg 31st May). Output used by \code{fill_calendar_data()}
}
