% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{create_overlapping_tiles}
\alias{create_overlapping_tiles}
\title{Create overlapping tile set}
\usage{
create_overlapping_tiles(template.r, overlap = 1000, sz = 10000)
}
\arguments{
\item{template.r}{= 1km land mask spatraster to inform whether land present}

\item{overlap}{= overlap in metres of each tile (x & y) - must be less than \code{sz}}

\item{sz}{= base size of tiles in metres. Where the remaining area is up to half a tile,
the last tile in each direction is extended (up to 1.5 x \code{sz}), otherwise an extra smaller
tile is added. If \code{sz} exceeds the area in one direction a single tile spans that direction.}
}
\value{
named list where "tile_extents" holds the terra extent of each tile and
"tile_land" is 'y' or 'n' indicating whether land cells within tile (based on template.r).
Returned list ordered by cols.
}
\description{
Create overlapping tile set
}
\examples{
r<-terra::rast(system.file("extdata/dtms/dtmf.tif",package="mesoclim"))
testtiles<-create_overlapping_tiles(r,overlap=200,sz=2000)
terra::plot(r)
for(t in testtiles$tile_extents) terra::plot(terra::vect(terra::ext(t)),add=TRUE)
testtiles<-create_overlapping_tiles(r,overlap=200,sz=20000)
}
\keyword{spatial}
