% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{create_parcel_list}
\alias{create_parcel_list}
\title{Calculates climate timeseries for vector parcels}
\usage{
create_parcel_list(
  climdata,
  parcels,
  id = "gid",
  input_names = c("tmax", "tmin", "swrad", "lwrad", "relhum", "pres", "prec",
    "windspeed"),
  output_names = c("tmax", "tmin", "swdown", "lwdown", "relhum", "pres", "prec",
    "windspeed"),
  roundings = c(2, 2, 1, 1, 1, 1, 2, 3)
)
}
\arguments{
\item{climdata}{\itemize{
\item list of SpatRasters of downscaled mesoclimate variables as output by \code{spatialdownscale}
}}

\item{parcels}{\itemize{
\item sf or vect describing polygons for which climate data is required
}}

\item{id}{\itemize{
\item string of variable name in parcels corresponding to parcel identification field
}}
}
\value{
nested list of climate variable timeseries by parcel.
}
\description{
Calculates timeseries of climate variables for each polygon in parcels using weighted means of overlapping grid cells.
}
\keyword{data}
\keyword{postprocess}
