% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{create_sst_data}
\alias{create_sst_data}
\title{Create UKCP sea surface temperature rast stack from ncdf files}
\usage{
create_sst_data(
  dir_data,
  startdate,
  enddate,
  aoi = NA,
  members = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15"),
  v = "SST"
)
}
\arguments{
\item{dir_data}{directory holding SST .nc files downloaded from ceda}

\item{startdate}{start date as POSIXlt}

\item{enddate}{end date as POSIXlt}

\item{aoi}{SpatRaster, SpatVector or sf object defining area of interest to crop data. NA no corpping occurs}

\item{members}{model members to be included}

\item{v}{= SST sea surface temperature}
}
\description{
Create UKCP sea surface temperature rast stack from ncdf files
}
