% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{create_ukcpsst_data}
\alias{create_ukcpsst_data}
\title{Create UKCP sea surface temperature rast stack from ncdf files.
Resamples and reprojects sea surface data to aoi as well as interpolating any missing coastal values.
Interpolation method = mean of all neighbouring cells.}
\usage{
create_ukcpsst_data(
  dir_data,
  startdate,
  enddate,
  aoi = NA,
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15"),
  v = "SST"
)
}
\arguments{
\item{dir_data}{directory holding SST .nc files downloaded from ceda}

\item{startdate}{start date as POSIXlt}

\item{enddate}{end date as POSIXlt}

\item{aoi}{SpatRaster, SpatVector or sf object defining area of interest to crop data. NA no cropping occurs}

\item{v}{= SST sea surface temperature variable name}

\item{members}{model members to be included}
}
\value{
Spatraster timeseris of sea surface temperatures resampled to aoi projection and resolution
}
\description{
Create UKCP sea surface temperature rast stack from ncdf files.
Resamples and reprojects sea surface data to aoi as well as interpolating any missing coastal values.
Interpolation method = mean of all neighbouring cells.
}
\examples{
 \dontrun{
 cedausr<-"your_user_name"
 cedapwd <- "your_password"
 startdate<-as.POSIXlt('2017/12/31')
 enddate<-as.POSIXlt('2018/12/31')
 member<-c('01')
 dir_out<-tempdir()
 download_ukcpsst(dir_out,startdate,enddate,member, cedausr,cedapwd)
 sst<-create_ukcpsst_data(dir_out,as.POSIXlt('2018/05/01'),as.POSIXlt('2018/05/31'),member=member)
 }
}
\keyword{preprocess}
\keyword{ukcp18}
