% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{daylength}
\alias{daylength}
\title{Calculates day length}
\usage{
daylength(julian, lat)
}
\arguments{
\item{julian}{\itemize{
\item astronomical julian day - as returned by .jday()
}}

\item{lat}{\itemize{
\item latitude (decimal degrees)
}}
}
\value{
Returns daylength in decimal hours (0 if 24 hour darkness, 24 if 24 hour daylight)
}
\description{
Calculates day length
}
\examples{
tme<-as.POSIXlt(seq(as.POSIXlt("2022-01-01"),as.POSIXlt("2022-06-30"),60*60*24*8))
jd<-sapply(tme,.jday)
dl<-daylength(jd,50)
plot(dl ~ jd, type = "l", lwd = 8, xlab = "Day", ylab = "Day length")
}
