% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{difprop}
\alias{difprop}
\title{Calculates the diffuse fraction from incoming shortwave radiation}
\usage{
difprop(rad, julian, localtime, lat, long, hourly = FALSE, merid = 0, dst = 0)
}
\arguments{
\item{rad}{a vector of incoming shortwave radiation values (W/m^2)}

\item{julian}{the Julian day as returned by \code{\link[=julday]{julday()}}}

\item{localtime}{a single numeric value representing local time (decimal hour, 24 hour clock)}

\item{lat}{a single numeric value representing the latitude of the location for which partitioned radiation is required (decimal degrees, -ve south of equator).}

\item{long}{a single numeric value representing the longitude of the location for which partitioned radiation is required (decimal degrees, -ve west of Greenwich meridian).}

\item{hourly}{specifies whether values of \code{rad} are hourly (see details).}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}
}
\value{
a vector of diffuse fractions (either \ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}} or \ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^{-2}}}).
}
\description{
\code{difprop} calculates proportion of incoming shortwave radiation that is diffuse radiation using the method of Skartveit et al. (1998) Solar Energy, 63: 173-183.
}
\details{
The method assumes the environment is snow free. Both overall cloud cover and heterogeneity in
cloud cover affect the diffuse fraction. Breaks in an extensive cloud deck may primarily
enhance the beam irradiance, whereas scattered clouds may enhance the diffuse irradiance and
leave the beam irradiance unaffected.  In consequence, if hourly data are available, an index
is applied to detect the presence of such variable/inhomogeneous clouds, based on variability
in radiation for each hour in question and values in the preceding and deciding hour.  If
hourly data are unavailable, an average variability is determined from radiation intensity.
}
\examples{
rad <- c(1:1352) # typical values of radiation in W/m^2
jd <- mesoclim:::.jday(as.POSIXlt("2022-06-21")) # julian day
dfr <- difprop(rad, jd, 12, 50, -5)
plot(dfr ~ rad, type = "l", lwd = 2,
xlab = expression(paste("Incoming shortwave radiation (", W*M^-2, ")")),
ylab = "Diffuse fraction")
}
\keyword{climate}
