% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.apply_cad}
\alias{.apply_cad}
\title{Calculate cold air drainage correction}
\usage{
.apply_cad(lrf, mu, st)
}
\arguments{
\item{lrf}{spatraster of lapse rates}

\item{mu}{spatraster of lapse rate multiplier output by\code{.cad_multiplier}.}

\item{st}{spatraster of suitability for cold air drainage as output by \code{.cad_conditions}}
}
\value{
spatraster of cold air drainage correction in deg C
}
\description{
Calculate cold air drainage correction
}
\keyword{internal}
