% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.basindelin}
\alias{.basindelin}
\title{Internal basin delineation function with option to merge by boundary}
\usage{
.basindelin(dtm, boundary = 0)
}
\description{
Internal basin delineation function with option to merge by boundary
}
