% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.basindelin_big}
\alias{.basindelin_big}
\title{Function used for delineating basins with big dtms}
\usage{
.basindelin_big(dtm, boundary = 0, tilesize = 100, plotprogress = FALSE)
}
\description{
Function used for delineating basins with big dtms
}
