% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{.blendmosaic}
\alias{.blendmosaic}
\title{blend two adjacent rasters that have overlap
called by \code{mosaicblend}}
\usage{
.blendmosaic(r1, r2)
}
\description{
blend two adjacent rasters that have overlap
called by \code{mosaicblend}
}
