% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.cad_multiplier}
\alias{.cad_multiplier}
\title{Calculate cold air drainage lapse rate multiplier}
\usage{
.cad_multiplier(dtmf, basins = NA, refhgt = 2)
}
\arguments{
\item{dtmf}{\itemize{
\item fine scale dtm spatraster
}}

\item{basins}{\itemize{
\item fine scale map of basins - if NA will calculate from the dtmf
}}

\item{refhgt}{\itemize{
\item Height of temperature measruements above ground
}}
}
\value{
spatRaster of lapse rate multiplier mu
}
\description{
Calculate cold air drainage lapse rate multiplier
}
\details{
calculates the lapse rate multiplier by the difference
in elevation between each point and the highest in the
basin.
}
\examples{
 dtmf<-terra::rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
plot(.cad_multiplier(dtmf, basins = NA, refhgt = 2))
}
\keyword{internal}
