% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.cadpotential}
\alias{.cadpotential}
\title{Calculate cold-air drainage potential (spatial)}
\usage{
.cadpotential(dtm, basins = NA, refhgt = 2)
}
\description{
Calculate cold-air drainage potential (spatial)
}
