% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{.change_rast_units}
\alias{.change_rast_units}
\title{Convert units and values of a SpatRast}
\usage{
.change_rast_units(r, to_unit)
}
\arguments{
\item{r}{SpatRaster with defined units}

\item{to_unit}{units of output SpatRaster}
}
\value{
SpatRaster in new units
}
\description{
Convert units and values of a SpatRast
}
\examples{
r<-rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
terra::units(r)<-'m'
r2<-.change_rast_units(r,'cm')
}
\keyword{internal}
