% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.clearskyraddaily}
\alias{.clearskyraddaily}
\title{Calculates average daily clear sky radiation over all pixels of a SpatRaster object and all days in tme}
\usage{
.clearskyraddaily(tme, r)
}
\arguments{
\item{tme}{\itemize{
\item a POSIXlt object of dates
}}

\item{r}{\itemize{
\item a terra::SpatRaster object
}}
}
\value{
a 3D array of expected daily clear sky radiation values (W/m**2)
}
\description{
Calculates average daily clear sky radiation over all pixels of a SpatRaster object and all days in tme
}
