% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.correctcoastal}
\alias{.correctcoastal}
\title{function for applying correction factor to coastal effects to account for
data resolution}
\usage{
.correctcoastal(r)
}
\description{
function for applying correction factor to coastal effects to account for
data resolution
}
