% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{.fill_calendar_data}
\alias{.fill_calendar_data}
\title{Converts 360-day yearly timeseries of UKCP18 raster data to actual calendar data}
\usage{
.fill_calendar_data(
  ukcp_r,
  real_dates,
  testplot = FALSE,
  plotdays = c(89:91, 242:244)
)
}
\arguments{
\item{ukcp_r}{SpatRaster timeseries stack loaded from UKCP18 .ncdf file}

\item{real_dates}{vector of valid date strings, as output by \code{.correct_ukcp_dates()} corresponding to the time period of 'ukcp_r'}

\item{testplot}{logical if TRUE plots example data of missing then interpolated days}

\item{plotdays}{the example days of years to plot DO NOT CHANGE}
}
\value{
a SpatRaster timeseries corresponding to actual calendar dates with invalid, no missing/empty days
}
\description{
Reassigns data and interpolates missing days of UKCP18 (360 day years) SpatRast data to create a new set of rasters corresponding to actual calendar dates.
}
\examples{
dir_ukcp<-system.file('extdata/ukcp18rcm',package='mesoclim')
nc<-file.path(dir_ukcp,"tasmax_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc")
ukdates<-.get_ukcp18_dates(nc)
real_dates<-.correct_ukcp_dates(ukdates)
r<-rast(nc,subds='tasmax')
r2<-.fill_calendar_data(r,real_dates,testplot=TRUE)
}
\keyword{internal}
