% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.latsfromr}
\alias{.latsfromr}
\title{Produces a matrix of latitudes form a terra::SpatRaster object
Inputs:
r - a terra::SpatRaster object
Returns a matrix of latidues}
\usage{
.latsfromr(r)
}
\description{
Produces a matrix of latitudes form a terra::SpatRaster object
Inputs:
r - a terra::SpatRaster object
Returns a matrix of latidues
}
