% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.lonsfromr}
\alias{.lonsfromr}
\title{Produces a matrix of longitudes form a terra::SpatRaster object
Inputs:
r - a terra::SpatRaster object
Returns a matrix of longitudes}
\usage{
.lonsfromr(r)
}
\description{
Produces a matrix of longitudes form a terra::SpatRaster object
Inputs:
r - a terra::SpatRaster object
Returns a matrix of longitudes
}
