% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{.lwup}
\alias{.lwup}
\title{Function to calculate LWup from temperature}
\usage{
.lwup(tc, sb = 5.67 * 10^-8, em = 0.97)
}
\arguments{
\item{tc}{temperature in C}

\item{sb}{stefan boltzman}

\item{em}{emissivity of surface c. 0.97 if not metal}
}
\value{
upward longwave radiation
}
\description{
Function to calculate LWup from temperature
}
\keyword{internal}
