% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{.multi_download}
\alias{.multi_download}
\title{ILya: Have changed this to an internal function, including calls to it.}
\usage{
.multi_download(f, url, dir_out, h)
}
\description{
ILya: Have changed this to an internal function, including calls to it.
}
