% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.saverast}
\alias{.saverast}
\title{writes out an array as a terra SpatRaster object}
\usage{
.saverast(a, template, path, varn)
}
\arguments{
\item{a}{\itemize{
\item a 3D array
}}

\item{path}{\itemize{
\item directory in which to save data
}}

\item{varn}{\itemize{
\item name of variable to be saved
}}

\item{description}{template - a SpatRaster object with the extent and coordinate reference system matching a}
}
\description{
writes out an array as a terra SpatRaster object
}
