% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{.swdown}
\alias{.swdown}
\title{Shortwave net to shortwave down}
\usage{
.swdown(swnet, cloud, dtmc, wsalbedo = 0.19, bsalbedo = 0.22, seaalb = 0.065)
}
\arguments{
\item{swnet}{Spatrast of net SW}

\item{cloud}{Spatrast of cld cover as \%}

\item{dtmc}{elevation (sea as NA) matching climate data}

\item{wsalbedo}{white sky albedo - can be a constant of fixed land albedo or a Spat raster of values.}

\item{bsalbedo}{as wsalbedo but black sky}

\item{seaalb}{sea albedo value to be used}
}
\value{
Spatraster timeseries of Downward shortwave radiation
}
\description{
Function to calculate SW down from SW net, cloud cover and white & black sky albedo
}
\examples{
#swdown_r<-.swdown(clim_list$rss,clim_list$clt,dtmc,wsalbedo,bsalbedo)
plot(swdown_r[[contrast_layers(swdown_r)]])
swdown_r<-.swdown(clim_list$rss,clim_list$clt,dtmc,NA,NA)
plot(swdown_r[[contrast_layers(swdown_r)]])
}
\keyword{internal}
