% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tempcad}
\alias{.tempcad}
\title{Downscale temperature with cold air drainage effects}
\usage{
.tempcad(climdata, dtmf, basins = NA, refhgt = 2)
}
\description{
Downscale temperature with cold air drainage effects
}
