% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tempcoastal}
\alias{.tempcoastal}
\title{Coastal temperature effects}
\usage{
.tempcoastal(
  tc,
  sstf,
  u2,
  wdir,
  dtmf,
  dtmm,
  dtmc,
  ndir = 32,
  smooth = 5,
  correct = TRUE
)
}
\arguments{
\item{tc}{\itemize{
\item downscaled temperature at same res as dtmf
}}

\item{sstf}{\itemize{
\item downscaled sea surface temperature to dtmf resolution and extent - no NA and timeseries must match tc
}}

\item{u2}{\itemize{
\item downscaled windspeed at temprature height
}}

\item{wdir}{\itemize{
\item wind direction (coarse resolution) - same value for all of dtmf extent will be used
}}

\item{dtmf}{\itemize{
\item fine dtm spatraster
}}

\item{dtmm}{\itemize{
\item fine-scale dtm covering wider area than dtmf (but same resolution!!)
}}

\item{dtmc}{\itemize{
\item coarse dtm spatraster matching resolution and extent of wdir
}}

\item{ndir}{\itemize{
\item number of directions to calculate for wind exposure
}}

\item{smooth}{\itemize{
\item number of cells to use for smoothing matrix
}}

\item{correct}{\itemize{
\item corrects each timestep  so that output mean temp of area matches input area mean temp
}}
}
\value{
Spatraster of temperature that includes coastal effect
}
\description{
Coastal temperature effects
}
\keyword{internal}
