% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tempelev}
\alias{.tempelev}
\title{Downscales temperature for elevation effects}
\usage{
.tempelev(tc, lrf, lrc, dtmf, dtmc = NA)
}
\description{
Downscales temperature for elevation effects
}
\keyword{internal}
