% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tmeinterp}
\alias{.tmeinterp}
\title{Temporal interpolation of Spatraster timeseries}
\usage{
.tmeinterp(r, tmein, tmeout)
}
\arguments{
\item{tmein}{= NA or Date or POSIX vector of datetimes of r - if NA assumes time(r) provides tmein values}

\item{tmeout}{= Date or POSIX vector of datetimes required}

\item{rin}{= 3D Spatraster}
}
\value{
Interpolated (linear) Spatraster stack of nlyrs=length(tmeout)
}
\description{
Temporal interpolation of Spatraster timeseries
}
\examples{
r <- terra::rast(ncols=5, nrows=5, vals=rep(1,25))
r<-c(r,r*2,r*3,r*4,r*5)
tmein<-as.Date(seq(17532,17652,30))
terra::time(r)<-tmein
tmeout<-as.Date(seq(17532,17651,1))
rout<-mesoclim:::.tmeinterp(r,NA,tmeout)
terra::plot(rout[[c(1,15,31,95,119,120)]])
}
\keyword{internal}
