% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.windz}
\alias{.windz}
\title{Calculates wind altitude coefficient in specified direction}
\usage{
.windz(dtm1, dtm2, dtmr, wdir)
}
\description{
Calculates wind altitude coefficient in specified direction
}
