% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_functions.R
\name{download_ancillary_era5}
\alias{download_ancillary_era5}
\title{Download ancillary era5 data (geopotential and lansea mask)}
\usage{
download_ancillary_era5(
  dir_out,
  area = c(61, -11.25, 48.75, 2.5),
  variables = c("geopotential", "land_sea_mask"),
  era5_user,
  era5_key
)
}
\arguments{
\item{dir_out}{\itemize{
\item directory to download files to
}}

\item{area}{\itemize{
\item lat lon of area
}}

\item{variables}{one or both of geopotential and land_sea_mask}

\item{era5_user}{CDS user name}

\item{era5_key}{CDS API token}
}
\value{
filepaths of outputs
}
\description{
Download ancillary era5 data (geopotential and lansea mask)
}
\details{
Downloads a single instance of ancillary dat for reanalysis-era5-single-levels as netcdf files.
}
\examples{
\dontrun{
dir_out <- tempdir()
cds_user<-"your_user_name"
cds_key<-"your_key_string"
download_ancillary_era5(dir_out,area = c(55,-3,49,3),variables =c("geopotential","land_sea_mask" ),era5_user=cds_user,era5_key=cds_key)
}
}
