% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_functions.R
\name{download_era5}
\alias{download_era5}
\title{Download ERA5 reanalysis data}
\usage{
download_era5(
  dir_out,
  file_out,
  year,
  area = c(61, -11.25, 48.75, 2.5),
  variables = c("10m u-component of wind", "10m v-component of wind",
    "2m dewpoint temperature", "2m temperature", "Mean sea level pressure",
    "Surface pressure", "Total precipitation", "Surface solar radiation downwards",
    "Total sky direct solar radiation at surface",
    "Mean surface downward short-wave radiation flux",
    "Mean surface direct short-wave radiation flux",
    "Mean surface downward long-wave radiation flux",
    "Mean surface net long-wave radiation flux"),
  era5_user,
  era5_key
)
}
\arguments{
\item{dir_out}{\itemize{
\item Local output directory
}}

\item{file_out}{\itemize{
\item Output filename
}}

\item{year}{\itemize{
\item year of data to retrieve
}}

\item{area}{\itemize{
\item bounding box with order c(ymax,xmin,ymin,xmax)
}}

\item{variables}{\itemize{
\item Long names of variables to retrieve
}}

\item{era5_user}{\itemize{
\item cds user name/email
}}

\item{era5_key}{\itemize{
\item cds api key
}}
}
\value{
filepath of output
}
\description{
Download ERA5 reanalysis data
}
\details{
Downloads a whole single year of reanalysis-era5-single-levels as netcdf file.
Variables chosen and size of area defaults match UK/Eire and variables required for microclimate modelling.
Limits on download file may restrict size of area and/or number of variables that can be downloaded to single file.
}
\examples{
\dontrun{
dir_out <- tempdir()
cds_user<-"your_user_name"
cds_key<-"your_key_string"
download_era5(dir_out,file_out='test.nc',year=2020,area = c(55,-3,49,3),variables =c("2m temperature"),era5_user=cds_user,era5_key=cds_key)
}
}
