% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_globalbedo}
\alias{download_globalbedo}
\title{Download 1km albedo data}
\usage{
download_globalbedo(
  dir_out,
  tiles = c("h17v03", "h17v04", "h18v03"),
  months = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"),
  years = seq(1998, 2011),
  url = "http://gws-access.jasmin.ac.uk",
  path = "/public/qa4ecv/albedo/netcdf_cf/1km/monthly"
)
}
\arguments{
\item{dir_out}{directory to which to save data}

\item{tiles}{vector of strings of the file name in the format 'h##v##'. Whole of the UK covered by default tiles.}

\item{year}{numeric vector indicating the years for which data are required}

\item{month}{string vector  indicating the month for which data are required}
}
\description{
The function \code{download_globalbedo} downloads monthly GlobAlbedo tiled
albedo data from Jasmin available for the years 1998-2011. Whole of the UK covered by default tiles. For further information see:
\url{http://www.globalbedo.org/index.php}
}
\details{
No username or password currently required.
}
