% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_haduk}
\alias{download_haduk}
\title{Downloads 1km gridded historic climata data for the UK}
\usage{
download_haduk(
  dir_out,
  startdate,
  enddate,
  varn = c("rainfall", "tasmax", "tasmin"),
  freq = c("day", "mon"),
  cedausr,
  cedapwd,
  access_token
)
}
\arguments{
\item{dir_out}{directory to which to save data}

\item{startdate}{\itemize{
\item POSIXlt value indicating start date for data required
}}

\item{enddate}{\itemize{
\item POSIXlt value indicating end date for data required
}}

\item{varn}{vector of variables required, one or more of: 'rainfall','tasmax','tasmin','hurs','psl','pv','sun','sfsWind'.
NOTE: Only \code{rainfall} (precipitation in mm), \code{tasmax}(maximum daily temperature, deg C) and
\code{tasmin} (minimum daily temperature, deg C) are available at daily time steps (all other monthly only).}

\item{freq}{\itemize{
\item string of frequency required - either 'day' or 'mon' (monthly)
}}

\item{cedausr}{ceda username string}

\item{cedapwd}{ceda password sting}

\item{access_token}{ceda access token as provided by \href{https://services-beta.ceda.ac.uk/account/token/?_ga=2.221117848.1874942177.1727205775-405001952.1710779653}{Ceda access token generator}}
}
\value{
names of downloaded files
}
\description{
The function \code{download_hadukdaily} downloads 1km HadUK-Grid Gridded
Climate Observation data on a 1km grid over the UK from CEDA.
}
\details{
Uses ceda dap service.
}
\examples{
\dontrun{
dir_out <- tempdir()
cedausr<-"your_user_name"
cedapwd <- "your_password"
access_token<-"your_token"
download_haduk(dir_out,as.POSIXlt('2018-01-01'),as.POSIXlt('2018-12-31'),varn=c('tasmax','tasmin'),freq='day', cedausr,cedapwd, access_token)
}
}
\keyword{download}
