% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_hadukdaily}
\alias{download_hadukdaily}
\title{Downloads 1km historic climata data for the UK}
\usage{
download_hadukdaily(dir_out, cedausr, cedapwd, year, month, varn)
}
\arguments{
\item{dir_out}{directory to which to save data}

\item{cedausr}{string of ceda username}

\item{cedapwd}{string of ceda username}

\item{year}{numeric value indicating the year for which data are required}

\item{month}{numeric value (1-12) indicating the month for which data are required}

\item{varn.}{Variable required. One of \code{rainfall} (precipitation in mm), \code{tasmax}
(maximum daily temperature, deg C) or \code{tasmin} (minimum daily temperature, deg C).}
}
\description{
The function \code{download_hadukdaily} downloads HadUK-Grid Gridded
Climate Observation data on a 1km grid over the UK from CEDA.
}
\details{
To obtain a username and password, first register at https://services.ceda.ac.uk/.
Your username will be the same as that given for your main CEDA account.
Your FTP password is separate from the password for your CEDA web account. If
you do not already have a password, or if you want to reset your FTP password
go to MyCEDA and click on the 'Configure FTP account' button.
}
