% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_ukcp18}
\alias{download_ukcp18}
\title{Download UKCP18 climate data}
\usage{
download_ukcp18(
  dir_out,
  startdate,
  enddate,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  rcp = c("rcp85", "rcp26"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  vars = c("clt", "hurs", "huss", "pr", "prsn", "psl", "rls", "rss", "tasmax", "tasmin",
    "uas", "vas"),
  download_dtm = FALSE,
  cedausr,
  cedapwd,
  access_token
)
}
\arguments{
\item{dir_out}{Output directory to which files are downloaded}

\item{startdate}{POSIXlt class defining start date of required timeseries}

\item{enddate}{POSIXlt class defining end date of required timeseries}

\item{collection}{text string defining UKCP18 collection, either 'land-gcm' or 'land-rcm'}

\item{domain}{text string defining UKCP18 domain, either 'uk' or 'eur'(land-rcm collection only) or 'global'}

\item{rcp}{text string iof RCP scenario to be downloaded}

\item{member}{vector of strings defining the climate model member to be downloaded. Available members vary between UKCP18 collections.}

\item{vars}{vector of strings corresponding to UKCP18 short variable names to download}

\item{download_dtm}{\itemize{
\item if TRUE and orography (elevation) available will also download to dir_out
}}

\item{cedausr}{ceda username string}

\item{cedapwd}{ceda password sting}

\item{access_token}{ceda access token as provided by \href{https://services-beta.ceda.ac.uk/account/token/?_ga=2.221117848.1874942177.1727205775-405001952.1710779653}{Ceda access token generator}}
}
\value{
downloads files to \code{dir_out}
}
\description{
Using parameters the function will downloaded from ftp.ceda.ac.uk all available UCKP18 files containing relevant data to the user defined output directory
}
\details{
start and end times are only used to identify decadal files containing data for this
time period - downloaded files are not cut to time period
}
\examples{
 \dontrun{
dir_ukcp <- tempdir()
cedausr<-"your_user_name"
cedapwd <- "your_password"
access_token<-"your_access_token"
download_ukcp18(dir_ukcp,as.POSIXlt('2018-05-01'),as.POSIXlt('2018-05-31'),'land-rcm','uk','rcp85',c('01'),c('tasmax','tasmin'),download_dtm=TRUE, cedausr,cedapwd,access_token)
}
}
\keyword{download}
\keyword{ukcp18}
