% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_ukcpsst}
\alias{download_ukcpsst}
\title{Download UKCP18rcm driven sea surface temperature predictions}
\usage{
download_ukcpsst(
  dir_out,
  startdate,
  enddate,
  modelruns = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  cedausr,
  cedapwd,
  cedaprot = "https://",
  cedaserv = "dap.ceda.ac.uk"
)
}
\arguments{
\item{dir_out}{\itemize{
\item local directory to which files written
}}

\item{startdate}{\itemize{
\item start date as POSIXlt
}}

\item{enddate}{\itemize{
\item end date as POSIXlt
}}

\item{modelruns}{\itemize{
\item vector of two character strings of model numbers
}}

\item{cedausr}{ceda username string}

\item{cedapwd}{ceda password sting}

\item{cedaprot}{\itemize{
\item string of ceda protocol, set to https DO NOT CHANGE
}}

\item{cedaserv}{\itemize{
\item string of the ceda server url, uses dap DO NOT CHANGE
}}
}
\value{
database detailing download details and success of all requested files
}
\description{
For description of the modelling used to produce sea surface temperatures then see:
\verb{https://www.metoffice.gov.uk/binaries/content/assets/metofficegovuk/pdf/research/ukcp/ukcp18-guidance-data-availability-access-and-formats.pdf}
}
\examples{
# cedausr<-"your_user_name"
# cedapwd <- "your_password"
# startdate<-as.POSIXlt('2017/12/31')
# enddate<-as.POSIXlt('2018/12/31')
# modelruns<-c('01')
# dir_out<-tempdir()
# download_ukcpsst(dir_out,startdate,enddate,modelruns, cedausr,cedapwd)
}
\keyword{download}
\keyword{ukcp18}
