% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtmf}
\alias{dtmf}
\title{A 100m grid resolution digital elevation dataset}
\format{
A PackedSpatRaster object with 197 rows and 214 columns
}
\source{
\url{http://www.tellusgb.ac.uk/}
}
\usage{
dtmf
}
\description{
A spatial dataset of elevations in (m) for the Lizard Peninsula, COrnwall, UK,
covering the area bounded by 160000, 181400, 11300, 30000  (xmin, xmax, ymin, ymax)
with the coordinate reference system OSGB36 / British National Grid (EPSG:27700)
}
\keyword{datasets}
