% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{era5sst}
\alias{era5sst}
\title{A 0.25 degree grid resolution dataset of sea-surface temperature data}
\format{
A PackedSpatRaster object with 8 rows, 8 columns and 744 layers
}
\source{
\url{https://cds.climate.copernicus.eu//}
}
\usage{
era5sst
}
\description{
A spatial dataset of hourly sea-surface temperatures for May 2018 for sea around
West Cornwall, UK, covering the area bounded by  -6.125, -4.125, 49.125, 51.125
(xmin, xmax, ymin, ymax) with the WGS84 lat long coordinate reference system (EPSG:4326)
}
\keyword{datasets}
