% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_functions.R
\name{era5toclimarray}
\alias{era5toclimarray}
\title{convert era4 ncdf4 file to format required for model}
\usage{
era5toclimarray(
  ncfile,
  dtmc,
  lsm = NA,
  aoi = NA,
  startdate,
  enddate,
  dtr_cor_fac = 1.285,
  toArrays = TRUE,
  zo = 10
)
}
\arguments{
\item{ncfile}{character vector containing the path and filename of the nc file}

\item{dtmc}{a SpatRaster object of ERA5 elevations of same or larger extent as aoi and same resolution as ncfile (see details)}

\item{lsm}{a SpatRaster object of ERA5 landsea mask of same or larger extent as aoi and same resolution as ncfile}

\item{aoi}{a  SpatRaster, sf or vect of the area to which outputs are cropped & reprojected - if NA dtmc used to define extent}

\item{startdate}{POSIXlt class defining start date of required timeseries}

\item{enddate}{POSIXlt class defining end date of required timeseries}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction of coastal grid cells. Default = 1.285, based on calibration against UK Met Office
observations. If set to zero, no correction is applied.}

\item{toArrays}{logical determining if climate data returned as list of arrays. If FALSE returns list of Spatrasters.}

\item{zo=}{height of wind speed above ground in metres to be output - defaults to ERA5 10m}
}
\value{
a list of the following:
\describe{
\item{dtmc}{Digital elevation of area (as Spatraster)}
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{windheight_m}{Height of windspeed data in metres above ground (as numeric)}
\item{tempheight_m}{Height of temperature data in metres above ground (as numeric)}
\item{temp}{Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swrad}{Total downward shortwave radiation (W/m^2)}
\item{difrad}{Downward diffuse radiation (W / m^2)}
\item{lwrad}{Total downward longwave radiation (W/m^2)}
\item{windspeed}{at 2m above ground (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
}
\description{
The function \code{era5toclimarray} converts data in a netCDF4 file returned
by mcera5 pkg function request_era5 to the correct formal required for subsequent modelling.
}
\details{
the model requires that input climate data are projected using a coordinate reference
system in which x and y are in metres. The output data are reprojected using the coordinate reference system and
extent of aoi (but SHOULD retain the approximate original grid resolution of the input climate data).
Suitable lsm can be derived from ERA5 landsea mask and dtmc can be derived from ERA5 geopotential ancillary data.
If \code{lsm} is not provided no coastal correction is applied. If \code{aoi} is not provided, \code{dtmc} and \code{lsm} are
cropped to the extent of the ERA5 data.
}
\examples{
 \dontrun{
ncfile<-system.file('extdata/era5raw/era5_sample.nc',package='mesoclim')
ancill<-terra::rotate(terra::rast(system.file('extdata/era5raw/era5_ancillary.nc',package='mesoclim')))
dtmc<-ancill$z / 9.80665
era5input<-era5toclimarray(ncfile, dtmc=dtmc, lsm=ancill$lsm,
                           startdate=as.POSIXlt('2020/01/01',tz="UTC"),
                           enddate=as.POSIXlt('2020/01/31',tz="UTC"))
checkinputs(era5input,'hour')
}
}
\keyword{era5}
\keyword{preprocess}
