% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{era5todaily}
\alias{era5todaily}
\title{Converts era5 hourly data to daily to enable bias correction to be applied to ukcp data}
\usage{
era5todaily(
  filein,
  pathout,
  landsea,
  elev,
  output = c("tifs", "singlenc", "none"),
  vars = c("tasmax", "tasmin", "sp", "psl", "hurs", "huss", "u10", "v10", "swrad",
    "difrad", "lwrad", "skyem", "pr")
)
}
\arguments{
\item{filein}{\itemize{
\item filename (including path ) of era5 nc file with required variables
}}

\item{pathout}{\itemize{
\item directory in which to save data
}}

\item{landsea}{\itemize{
\item a landsea raster object of land fractions that must match the extent (or less) and crs of the era5 data
era5 data will be cropped to lsm extent
}}

\item{elev}{\itemize{
\item a spatRaster of surface elevations to correct surface pressure to sea level pressure
}}

\item{output}{\itemize{
\item character of value tifs, singlenc or none - determines type of files (if any) written
}}

\item{vars}{\itemize{
\item variables for which daily values calculated
}}
}
\value{
list of daily spatRasters of each variable requested. Also write .tiff or .nc files if requested
The function saves individual terra SpatRaster objects to disk (one for each variable)
}
\description{
Converts era5 hourly data to daily to enable bias correction to be applied to ukcp data
}
\examples{
 \dontrun{
vars<-c('psl','hurs','u10','v10','swrad','difrad','lwrad')
daily<-.era5todaily(filein=file.path(dir_era5,f),tempdir(),landsea=lsm,output='none',vars)
}
}
\keyword{internal}
