% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{find_ukcp_decade}
\alias{find_ukcp_decade}
\title{Find UKCP18 decades containing timeseries of data}
\usage{
find_ukcp_decade(collection = c("land-gcm", "land-rcm"), startdate, enddate)
}
\arguments{
\item{collection}{UKCP18 collection, either global ('land-gcm') or regional ('land-rcm'), as file naming conventions vary between each collection}

\item{startdate}{POSIXlt class defining starting date}

\item{enddate}{POSIXlt class defining starting date}
}
\value{
a vector of strings corresponding to the decade part of UKCP18 files containing the entire requested timeseries
}
\description{
Returns UKCP18 decade text (used in file names) containing required data defined by start and end dates
}
\examples{
find_ukcp_decade('land-rcm',as.POSIXlt("2030/01/01"),as.POSIXlt("2039/12/31"))

}
