% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{flowacc}
\alias{flowacc}
\title{Calculates accumulated flow}
\usage{
flowacc(dtm, basins = NA)
}
\arguments{
\item{dtm}{a SpatRast elevations (m).}

\item{basins}{optionally a SpatRast of basins numbered as integers (see details).}
}
\value{
a SpatRast of accumulated flow (number of cells)
}
\description{
\code{flowacc} calculates accumulated flow(used to model cold air drainage)
}
\details{
Accumulated flow is expressed in terms of number of cells. If \code{basins}
is provided, accumulated flow to any cell within a basin can only occur from
other cells within that basin.
}
\examples{
r<-rast(matrix(c(25,24,21,24,25,15,14,11,14,15,5,4,1,4,5,15,14,11,14,15,25,24,21,24,25),nrow=5, ncol=5))
plot(log(flowacc(r)))
}
\keyword{spatial}
