% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trees_functions.R
\name{get_dtmm}
\alias{get_dtmm}
\title{Calculate medium resolution dtm from aoi, coarse (ukcp) dtmc and uk wide fine scale dtmuk}
\usage{
get_dtmm(aoi, dtmc, dtmuk, basepath = "")
}
\arguments{
\item{aoi}{\itemize{
\item area of interest spatvector, raster or sf
}}

\item{dtmc}{\itemize{
\item coarse dtm spatraster
}}

\item{dtmuk}{\itemize{
\item uk wide fine resolution dtm spatraster
}}

\item{basepath}{\itemize{
\item "" when using on jasmin
}}
}
\value{
spatraster object
}
\description{
Calculate medium resolution dtm from aoi, coarse (ukcp) dtmc and uk wide fine scale dtmuk
}
\examples{
\dontrun{
dtmm<-get_dtmm(aoi,dtmc,dtmuk,basepath=ceda_basepath)
}
}
\keyword{preprocess}
