% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_nearest_val}
\alias{get_nearest_val}
\title{Find nearest non-NA cell values to a point}
\usage{
get_nearest_val(r, pt)
}
\arguments{
\item{r}{spatraster of climate data}

\item{pt}{vect point}
}
\value{
Nearest non-NA cell values to pt of length nlyr(r)
}
\description{
Find nearest non-NA cell values to a point
}
