% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{get_ukcp18_dates}
\alias{get_ukcp18_dates}
\title{Gets dates of UKCP18 ncdf file}
\usage{
get_ukcp18_dates(ncfile)
}
\arguments{
\item{ncfile}{filename of UKCP18 ncdf data file}
}
\value{
a vector of date strings in the form 'yyy-dd-mm'
}
\description{
The function \code{era5toclimarray} converts data in a netCDF4 file returned
by \code{\link[mcera5:request_era5]{mcera5::request_era5()}} to the correct formal required for subsequent modelling.
}
\details{
Creates a time series from the ncdf file time variable which is not correctly read by R terra package.
UKCP18 time values expressed as hours since 1/1/1970 12.00. Output used by function \code{correct_ukcp_dates()}
}
