% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trees_functions.R
\name{get_ukcp_dtm}
\alias{get_ukcp_dtm}
\title{Get ukcp18 RCM 12km DTM from ceda archive}
\usage{
get_ukcp_dtm(aoi, basepath = "")
}
\arguments{
\item{aoi}{\itemize{
\item used to cut dtm
}}

\item{basepath}{\itemize{
\item "" when using on jasmin
}}
}
\value{
spatrast object cropped to extended aoi
}
\description{
Get ukcp18 RCM 12km DTM from ceda archive
}
\examples{
\dontrun{
dtmc<-get_ukcp_dtm(aoi, basepath=ceda_basepath)
}
}
\keyword{jasmin}
