% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{hum_dailytohourly}
\alias{hum_dailytohourly}
\title{Daily to hourly relative humidity}
\usage{
hum_dailytohourly(
  relhum,
  tasmin,
  tasmax,
  temph,
  psl,
  presh,
  tme,
  relmin = 2,
  adjust = TRUE
)
}
\arguments{
\item{relhum}{\itemize{
\item an array of daily relative humidity values (\%)
}}

\item{tasmin}{\itemize{
\item an array of daily minimum temperature values (deg C)
}}

\item{tasmax}{\itemize{
\item an array of daily maximum temperature values (deg C)
}}

\item{temph}{\itemize{
\item an array of hourly temperature values (deg C)
}}

\item{psl}{\itemize{
\item an array of daily sea-level pressure values(kPa)
}}

\item{presh}{\itemize{
\item an aray of hourly surface-level pressure values (kPa)
}}

\item{tme}{\itemize{
\item POSIXlt object of dates corresponding to radsw
}}

\item{relmin}{\itemize{
\item minimum possible relative humidity value
}}

\item{adjust}{\itemize{
\item optional logical which if TRUE ensures that, after interpolation, returned
hourly values, when averaged to daily, match the input
}}
}
\value{
an array of hourly relative humidity values (\%)
}
\description{
Daily to hourly relative humidity
}
\details{
#' Owing to the strong dependence of relative humidity on temperature, the
resulting diurnal patterns, prior to interpolation, relative humidity is first
converted to specific humidity using tasmin, tasmax and psl. After interpolation,
the data are back-converted to relative humidity, using temph and presh.
~~ Function should spline interpolate vapour pressure and use temperature cycle
Derives an array of hourly relative humidity values from daily values.
}
\examples{
\dontrun{
climdata<- read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
# Get hourly variables required for rel hum downscaling
dtmc<-climdata$dtm
temph<-as.array(hourlytemp(mesoclim:::.rast(climdata$tmin,dtmc), mesoclim:::.rast(climdata$tmax,dtmc), climdata$tme, 50, -2.5))
presh<-pres_dailytohourly(climdata$pres,climdata$tme)
# Get sea level pressure - MAKE THIS A FUNCTION!!
psl_r<-mesoclim:::.rast(climdata$pres,dtmc) / (((293-0.0065*ifel(is.na(dtmc),0,dtmc))/293)^5.26)
psl<-mesoclim:::.rta(psl_r,length(climdata$tme))
humh<-hum_dailytohourly(climdata$relhum, climdata$tmin, climdata$tmax, temph, psl, presh, climdata$tme)
plot_q_layers(mesoclim:::.rast(humh,dtmc))
}
}
\keyword{temporal}
