% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invls_calc}
\alias{invls_calc}
\title{This function returns a numeric matrix identifying
matching values from a raster that intersect with
a set of points}
\usage{
invls_calc(
  lsm,
  resolution,
  xmin,
  ymax,
  s,
  direction,
  slr,
  slr_xmin,
  slr_xmax,
  slr_ymin,
  slr_ymax
)
}
\description{
This function returns a numeric matrix identifying
matching values from a raster that intersect with
a set of points
}
\keyword{biascorrect}
