% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{landfill_climdata}
\alias{landfill_climdata}
\title{Fill coastal data}
\usage{
landfill_climdata(climdata, landsea_true, landsea_climdata = NA)
}
\arguments{
\item{climdata}{\itemize{
\item spatraster of climate data
}}

\item{landsea_true}{\itemize{
\item true landsea mask
}}

\item{landsea_climdata}{\itemize{
\item landsea mask of climdata
}}
}
\value{
Climdata to cover all land in landsea_true
}
\description{
Fill coastal data
}
