% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsea}
\alias{landsea}
\title{A 100m grid resolution landsea mask}
\format{
A PackedSpatRaster object with 500 rows and 500 columns
}
\source{
\url{https://cds.climate.copernicus.eu//}
}
\usage{
landsea
}
\description{
A spatial landsea mask for West Cornwall, UK, covering the area bounded by  145000,
195000, -9000, 41000 (xmin, xmax, ymin, ymax) with the with the coordinate reference
system OSGB36 / British National Grid (EPSG:27700)
}
\keyword{datasets}
