% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{lapserate}
\alias{lapserate}
\title{Calculate Lapse rates via vapour pressure}
\usage{
lapserate(tc, rh = NA, pk = NA)
}
\arguments{
\item{tc}{= temperature (deg C) as vector, array or spatraster}

\item{pk}{= atmospheric pressure (kPa) as vector, array or spatraster}

\item{ea}{= vapour pressure (kPa) as vector, array or spatraster}
}
\value{
lapse rates in same format as tc
}
\description{
Calculate Lapse rates via vapour pressure
}
\details{
Alternatively a value of 0.005 can be used if lacking pk or relhum data
}
\examples{
climdata<-read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
ea<-converthumidity(climdata$relhum,tc=climdata$temp , pk=climdata$pres)
lr<-lapserate(climdata$temp,ea,climdata$pres)
terra::plot(mesoclim:::.rast(lr,climdata$dtm)[[1]])
}
\keyword{climate}
\keyword{spatial}
