% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{multi_download}
\alias{multi_download}
\title{Carries out multiple downloads from ceda - called by download_ukcp18}
\usage{
multi_download(f, url, dir_out, h)
}
\arguments{
\item{f}{filename of file to be downloaded}

\item{url}{complete url of file to be downloaded}

\item{dir_out}{output directory of downloaded file}

\item{h}{curl handle for file download}
}
\value{
downloads file to \code{file.path(dir_out,f)}
}
\description{
Carries out multiple downloads from ceda - called by download_ukcp18
}
