% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect.R
\name{precipcorrect}
\alias{precipcorrect}
\title{Applies bias correction to precipitation data}
\usage{
precipcorrect(hist_obs, hist_mod, fut_mod, rangelims = NA, rangelim = 2)
}
\arguments{
\item{hist_obs}{a stacked SpatRast of precipitation data.}

\item{hist_mod}{a stacked SpatRast of modelled precipitation data for
for the same period and spatial area as \code{hist_obs}. I.e \code{dim(hist_mod)} must equal
\code{dim(hist_obs)}}

\item{fut_mod}{a stacked SpatRast of modelled precipitation data for
for e.g. a future period to which corrections are applied. Must have the same
x and y dims as \code{hist_obs} and \code{hist_mod}, but can cover a shorter or
longer time period.}

\item{rangelim}{fractional amount above the maximum values of precipitation
in \code{hist_obs} in any given cell by which values in the returned dataset can
attain. I.e. if set to 2, values in any given grid cell of the returned dataset
cannot exceed twice the maximum amount in the corresponding grid cell in
\code{hist_obs}.If not supplied (the default), returned values are unbounded.}
}
\value{
a SpatRast of bias corrected precipitation.
}
\description{
The function \code{biascorrect} applies bias correction to spatial
climate dataset
}
\seealso{
\code{\link[=biascorrect]{biascorrect()}} for applying corrections to other climate variables.
}
