% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect_functions.R
\name{precipcorrect_apply}
\alias{precipcorrect_apply}
\title{Applies bias correction model to precipitation data}
\usage{
precipcorrect_apply(fut_mod, biasmods, prec_thold = 0.01)
}
\arguments{
\item{fut_mod}{a stacked SpatRast of modelled precipitation data for
for e.g. a future period to which corrections are applied. reprojected and
resampled to match data held in \code{mod_out} if necessary.}

\item{biasmods}{list containing two wrapped SpatRasters derived using \code{\link[=precipcorrect]{precipcorrect()}}
representing the amount by which to adjust rainfall totals and rainfall days,
which is used as a basis for adjusting individual rainfall events.}

\item{prec_thold}{numeric value (mm per day) below which precipitation is taken to be zero}
}
\value{
a SpatRaster matching \code{fut_mod} with climate data bias corrected.
}
\description{
The function \code{precipcorrect_apply} applies a set of bias corrections
derived using \code{precip_correct} to a spatial climate dataset
}
\details{
The \code{precipcorrect_apply} function allows the same bias correction to
be applied to multiple datasets, spanning e.g. several future time periods or
climate scenarious. The spatial extent of \code{fut_mod} must at least partially
overlap with that of \code{hist_obs} when running \code{precipcorrect} as corrected values are only returned for the overlapping area
}
\seealso{
\code{\link[=precipcorrect]{precipcorrect()}} for deriving \code{biasmods} and \code{biascorrect_apply} for
applying corrections to other climate variables
}
\keyword{biascorrect}
