% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{pres_dailytohourly}
\alias{pres_dailytohourly}
\title{Daily to hourly atmospheric pressure}
\usage{
pres_dailytohourly(pres, tme = NA, adjust = TRUE)
}
\arguments{
\item{pres}{\itemize{
\item an array of daily mean surface-level pressure values (kPa)
}}

\item{tme}{\itemize{
\item POSIXlt object of dates corresponding to radsw
}}

\item{adjust}{\itemize{
\item optional logical which if TRUE ensures that, after interpolation, returned
hourly values, when averaged to daily, match the input
}}
}
\value{
an array of hourly surface-level pressure values (kPa)
}
\description{
Derives an array of hourly surface-level pressure values from an array of daily
surface level pressure values.
}
\details{
Note that, as in the era5 dataste, these are surface level pressures in kPa corrsponding to
the elevation of each geographic location and not sea=level pressures. The
conversion to millbars is mb = 10 kPa. The conversion to sea-level pressure (Psl) can be
derived by reversing the equation Psl = pres*((293-0.0065z)/293)**5.26
}
\examples{
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as vector =========================== #
# ========================================================================= #
# Derive daily tmx and tmn from hourly era5 data
pres <- c(99,96,94,92,100)
tme <- as.POSIXlt(c(0:4) * 3600 * 24, origin = "2018-05-01", tz = "UTC")
hrpres <- pres_dailytohourly(pres,tme)
hr_tme<-as.POSIXlt(unlist(lapply(tme,FUN=function(x) x+(60*60*c(0:23)) )))
# Plot results to compare
matplot(x=as.numeric(hr_tme),y=hrpres, type = "l", lty = 1)
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as SpatRaster ======================= #
# ========================================================================= #
climdaily<- read_climdata(mesoclim::ukcpinput)
hrpres<-pres_dailytohourly(climdaily$pres)
# Plot results for one cell
cell_pres<-t(extract(hrpres,matrix(c(175000,40000),ncol=2)))
matplot(as_datetime(time(cell_pres)),cell_pres, type = "l", lty = 1)
}
\keyword{temporal}
