% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{skyem_dailytohourly}
\alias{skyem_dailytohourly}
\title{Daily to hourly downward longwave radiation}
\usage{
skyem_dailytohourly(skyem, tme, adjust = TRUE)
}
\arguments{
\item{skyem}{\itemize{
\item an array of daily mean sky-emissivity values values (0-1)
}}

\item{tme}{\itemize{
\item POSIXlt object of dates corresponding to radsw
}}

\item{adjust}{\itemize{
\item optional logical which if TRUE ensures that, after interpolation, returned
hourly values, when averaged to daily, match the input
}}
}
\value{
an array of hourly sky-emissivity values values (0-1)
}
\description{
Derives an array of hourly effective sky-emissivity values.
}
\details{
NB - more consistent to code this as downward longwave, but will essentially
do the calaculations in the function below, but with temperature as an additional
Effective sky emissvity can be used to calaculate downward longwave radiation (Lwd).
The formula is Lwd = skyem * Lwu where Lwu is upward longwave radiation given
by Lwu=0.97\emph{5.67}10**-8*(tc+273.15). Here tc is average surface temperature (deg C))
but an adequate approximation is derived if subtited by air temperature.
}
\examples{
lwrad<-skyem_dailytohourly(array(rep(c(0.2,0.2,0.2,0.8,0.8,0.8),2),c(3,3,2)),as.POSIXlt(seq(1,(48*60*60),(24*60*60))))
}
\keyword{temporal}
