% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_functions.R
\name{spatial_interpNA}
\alias{spatial_interpNA}
\title{Spatially interpolate missing values in Spatraster}
\usage{
spatial_interpNA(r)
}
\arguments{
\item{r}{2 or 3D Spatraster}
}
\value{
Spatraster of same dim as r but without NA
}
\description{
Spatially interpolate missing values in Spatraster
}
\examples{
r <- rast(ncols=5, nrows=5, vals=rep(c(1,2,NA,4,5),5))
rout<-spatial_interpNA(r)
plot(c(r,rout))
}
