% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{spatialdownscale}
\alias{spatialdownscale}
\title{Spatially downscale all climate variables}
\usage{
spatialdownscale(
  climdata,
  sst,
  dtmf,
  dtmm = NA,
  basins = NA,
  wca = NA,
  skyview = NA,
  horizon = NA,
  cad = TRUE,
  coastal = TRUE,
  thgto = 2,
  whgto = 2,
  include_tmean = FALSE,
  rhmin = 20,
  pksealevel = TRUE,
  patchsim = FALSE,
  terrainshade = TRUE,
  precipmethod = "Elev",
  fast = TRUE,
  noraincut = 0,
  toArrays = FALSE,
  cex = NA
)
}
\arguments{
\item{climdata}{a \code{climdata} model object containing climate data of the same format as \code{ukcp18toclimarray}}

\item{sst}{a SpatRast of sea-surface temperature data (deg C) that overlaps with climdata$tme}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf. Needed for wind downscaling, and for coastal effects if \code{cex} isn't supplied
(see \code{\link[=calculate_coastalexposure]{calculate_coastalexposure()}}).}

\item{basins}{optionally, a fine-resolution SpatRast of basins as returned by \code{\link[=basindelin]{basindelin()}}
matching the coordinate reference system and extent of \code{dtmf}. Calculated if
not supplied.}

\item{wca}{optionally, an array of wind shelter coeffs in 8 directions
matching the x,y dimensions of \code{dtmf} as calculated by \code{calculate_windcoeffs}. Calculated if
not supplied.}

\item{cad}{optional logical indicating whether to calculate cold-air drainage effects}

\item{coastal}{optional logical indicating whether to calculate coastal effects}

\item{thgto}{height above ground of temperature output.}

\item{whgto}{height above ground of wind speed output.}

\item{include_tmean}{if TRUE and daily data will output mean daily temperature based on hourly downsccaling}

\item{rhmin}{minimum relative humidity (set to avoid relative humidity dropping too low
during down-scaling). Default 20 (percent).}

\item{pksealevel}{optional logical indicating whether input pressure data represent
sea-level pressure (default TRUE).}

\item{terrainshade}{optional logical indicating whether to account for terrain shading
during shortwave radiation downscale.}

\item{precipmethod}{One of \code{Tps} or \code{Elev} indicating whether to account for
elevation effects using a Thin-plate spline model or en emprical adjustment
(see details under \code{\link[=precipdownscale]{precipdownscale()}})}

\item{fast}{optional logical indicating whether to use fast Thin-plate spline
down-scaling (see details under \code{\link[=precipdownscale]{precipdownscale()}}).}

\item{noraincut}{numeric value below which low precipitation amounts are set to
zero (see details under \code{\link[=precipdownscale]{precipdownscale()}}).}

\item{toArrays}{if TRUE climate variables are output as 3D arrays if FALSE as Spatarasters}

\item{cex}{optionally, coastal exposure matching \code{dtmf} as returned by \code{\link[=calculate_coastalexposure]{calculate_coastalexposure()}}.
Calculated from \code{dtmf} and \code{dtmm} if not supplied and \code{coastal = TRUE}.}

\item{pathsim}{optional logical indicating whether to simulate cloud cover and
rainfall patchiness see details under \code{\link[=swdownscale]{swdownscale()}} and \code{\link[=precipdownscale]{precipdownscale()}}.}

\item{difr}{optional logical indicating whether to return diffuse radiation (
default TRUE).}
}
\value{
a list of multi-layer SpatRast or 3D arrays of downscaled climate variables and metadata as follows:
\describe{
\item{dtm}{Digital elevation of downscaled area in metres (as Spatraster)}
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{windheight_m}{Height of windspeed data in metres above ground (as numeric)}
\item{tempheight_m}{Height of temperature data in metres above ground (as numeric)}
\item{temp}{Temperature (deg C) - is replaced with \code{tmax} and \code{tmin} if input data is daily and \code{tmean} if \code{include_tmean}=TRUE.}
\item{relhum}{Relative humidity (Percentage)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swrad}{Total downward shortwave radiation (W/m^2)}
\item{difrad}{Downward diffuse radiation (W / m^2) - not returned unless \code{terrainshade = TRUE}}
\item{lwrad}{Total downward longwave radiation (W/m^2)}
\item{windspeed}{at 2m (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
}
\description{
Spatially downscales coarse-resolution climate data
}
\examples{
 dtmf<-terra::rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
 dtmm<-terra::rast(system.file('extdata/dtms/dtmm.tif',package='mesoclim'))
 climdata<-read_climdata(mesoclim::ukcpinput)
 sst<-terra::unwrap(mesoclim::ukcp18sst)
 mesodat<-spatialdownscale(climdata, sst, dtmf, dtmm, include_tmean=TRUE,noraincut=0.01)
 for(n in 5:length(mesodat)) terra::plot(mesodat[[n]][[12]],main=paste(names(mesodat)[n],mesodat$tme[n]))
}
\keyword{spatial}
