% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{SSTinterpolate}
\alias{SSTinterpolate}
\title{temporally interpolate sea-surface temperature data to hourly}
\usage{
SSTinterpolate(SST, tmein, tmeout)
}
\arguments{
\item{SST}{a SpatRast of hourly, daily or e.g. monthly sea-surface tmeperature data}

\item{tmein}{POSIXlt object of times corresponding to each layer of \code{SST}}

\item{tmeout}{POSIXlt object of times for which output sea-surface temperature data are needed (hourly)}
}
\value{
a SpatRast of sea-surface temperature data wiht the same units as \code{SST} (usually deg C).
}
\description{
The function \code{SSTinterpolate} spatially infills missing spatial spatial
sea-surface temperature data and then temporally interpolates data to hourly if not
already hourly.
}
\details{
Missing grid cells are interpolated using zoo::na.approx. Output sea-surface
temperature data are temporally interpolated to daily values, so the duration of
\code{tmeout} must be at least that of \code{tmeout}. Hourly values are returned by replicating
all values within a day as sea-surface temperatures tend to fluctuate slowly.
}
