% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{subset_climdata}
\alias{subset_climdata}
\title{Subset climate data by date}
\usage{
subset_climdata(climdata, sdatetime, edatetime)
}
\arguments{
\item{climdata}{\itemize{
\item list of spatraster/arrays and associated climate data
}}

\item{sdatetime}{\itemize{
\item start date time as POSIXlt or Date
}}

\item{edatetime}{\itemize{
\item end date time as POSIXlt or Date
}}
}
\value{
list of same format as \code{climdata} but retricted to times between \code{sdatetime} and \code{edatetime}
}
\description{
Subset climate data by date
}
\examples{
climdata<-read_climdata(mesoclim::ukcpinput)
start<-as.Date("2018/05/01")
end<-as.Date("2018/05/07")
weekdata<-subset_climdata(climdata,start,end)
}
