% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{temp_dailytohourly}
\alias{temp_dailytohourly}
\title{Derives hourly temperature from daily data}
\usage{
temp_dailytohourly(tmn, tmx, tme = NA, lat = NA, long = NA, srte = 0.09)
}
\arguments{
\item{tmn}{a vector or SpatRaster of daily minimum temperatures (deg C)}

\item{tmx}{a vector or SpatRaster of daily maximum temperatures (deg C)}

\item{tme}{POSIXlt object of dates. Only required if \code{tmn} and
\code{tmx} are vectors.}

\item{lat}{latitude of location (decimal degrees). Only required if \code{tmn} and
\code{tmx} are vectors}

\item{srte}{a parameter controlling speed of decay of night time temperatures (see details)}

\item{lon}{longitude of location (decimal degrees). Only required if \code{tmn} and
\code{tmx} are vectors}
}
\description{
Generates a plausible vector or SpatRaster of hourly temperatures
from daily data.
}
\details{
Daytime temperatures are assumed to follow a sine curve with a peak a
short while after solar noon. After dusk, the temperatures are assumed to decay
exponentially reaching a minimum at dawn. The day in which tmx and tmn fall is
assumed to match UTC days. The parameter \code{stre} controls the speed of decay of
night time temperatures with time. A value of zero ensures values drop to minimum
at dawn the following day, but trial and error indicates in most circumstances
temperatures decay faster than this. The default value of 0.09 is an optimal
value derived using ERA5 data for western Europe, but performs reasonably well
globally
}
\examples{
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as vector =========================== #
# ========================================================================= #
# Derive daily tmx and tmn from hourly era5 data
tmx <- c(15,15,18,20,12)
tmn <- c(10,9,12,12,6)
tme <- as.POSIXlt(c(0:4) * 3600 * 24, origin = "2018-05-01", tz = "UTC")
# Use interpolation function
hr_temp <- temp_dailytohourly(tmn, tmx, tme, 50, -5)
hr_tme<-as.POSIXlt(unlist(lapply(tme,FUN=function(x) x+(60*60*c(0:23)) )))
# Plot results to compare
matplot(x=as.numeric(hr_tme),y=hr_temp, type = "l", lty = 1)
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as SpatRaster ======================= #
# ========================================================================= #
climdaily<- read_climdata(mesoclim::ukcpinput)
hrtemps<-temp_dailytohourly(climdaily$tmin, climdaily$tmax, srte = 0.09)
# Plot results for one cell
cell_temp<-t(terra::extract(hrtemps,matrix(c(175000,40000),ncol=2)))
matplot(lubridate::as_datetime(terra::time(hrtemps)),cell_temp, type = "l", lty = 1)
}
\keyword{temporal}
