% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{tempcoastal}
\alias{tempcoastal}
\title{Downscale temperature with coastal effects}
\usage{
tempcoastal(tc, SST, u2, wdir, dtmf, dtmm, dtmc)
}
\arguments{
\item{tc}{an array of hourly temperature data in deg C (either coarse-resolution,
matching \code{dtmc} or fine-scale matching \code{dtmf})}

\item{SST}{a SpatRast of hourly sea-surface temperature data (deg C) with no NAs, as
returned by \code{\link[=SSTinterpolate]{SSTinterpolate()}}}

\item{u2}{a SpatRast of high-resolution wind speeds 2 m above ground as returned
by \code{\link[=winddownscale]{winddownscale()}}}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf (see details)}

\item{dtmc}{a coarse-resolution SpatRast of elevations matching
the resolution, extent and coordinate reference system of \code{wspeed}.}
}
\value{
a SpatRast of temperature differences owing to cold-air drainage (deg C).
}
\description{
Downscales an array of temperature data adjusting for coastal effects
}
\examples{
tc<-era5data$climarray$temp
tme<-as.POSIXlt(era5data$tme,tz="UTC")
SST<-SSTinterpolate(rast(era5sst),tme,tme)
wspeed<-era5data$climarray$windspeed
wdir<-era5data$climarray$winddir
dtmc<-rast(era5data$dtmc)
# Downscale wind - takes a few seconds
u2<-winddownscale(wspeed, wdir, rast(dtmf), rast(dtmm), dtmc, uz = 2)
# Apply coastal effects (takes a few seconds)
tcf<-tempcoastal(tc,SST,u2,wdir,rast(dtmf),rast(dtmm),dtmc)
par(mfrow=c(2,1))
plot(tcf[[1]])
plot(tcf[[12]])
}
