% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_functions.R
\name{time_interp}
\alias{time_interp}
\title{Temporal interpolation of missing times in Spatraster
Works for any time step not just daily or hourly eg monthly SST}
\usage{
time_interp(rin, tmeout, tmein = NA)
}
\arguments{
\item{rin}{= 3D Spatraster}

\item{tmeout}{= Date or POSIX vector of datetimes required}

\item{tmein}{= Date or POSIX vector of datetimes of rin - if NA assumes time(rin) provides valules}
}
\value{
Spatraster stack of nlyrs=length(tmeout)
}
\description{
Temporal interpolation of missing times in Spatraster
Works for any time step not just daily or hourly eg monthly SST
}
\examples{
r <- rast(ncols=5, nrows=5, vals=rep(1,25))
rin<-c(r,r*2,r*3,r*4,r*5)
tmein<-as.Date(seq(17532,17652,30))
terra::time(rin)<-tmein
tmeout<-as.Date(seq(17532,17651,1))
rout<-time_interp(rin,tmeout, NA)
plot(rout[[c(1,15,31,95,119,120)]])
}
