% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukcp18aus}
\alias{ukcp18aus}
\title{A list of ~60km resolution UKCP18 global climate data (global domain)}
\format{
a list with the following elements:
\describe{
\item{tme}{POSIXlt object of dates and times}
\item{climarray}{a list of arrays of climate variables with the following variables:}
\item{windheight_m}{numeric value in metres of wind height above ground}
\item{tempheight_m}{numeric value in metres of temperature height above ground}
\describe{
\item{tmax}{Maximum Daily Temperature (deg C)}
\item{tmin}{Minimum Daily Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{spechum}{Specific humidity}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swnet}{Net downward shortwave radiation (W/m^2)}
\item{lwnet}{Net downward longwave radiation (W/m^2)}
\item{windspeed at 2m (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
\item{dtmc}{a wrapped SpatRast object of elevations (m) matching the extent and resolution of climate data}
}
}
\source{
\url{ftp.ceda.ac.uk}
}
\usage{
ukcp18aus
}
\description{
A list of hourly UKCP18 regional climata data for May 2018 for East Australia UK as
returned by \link{ukcp18toclimarray}  where collection='land-gcm'& domain='global'.
Lon res = 0.833 deg; Lat res = 0.55 deg
}
\keyword{datasets}
