% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukcp18lookup}
\alias{ukcp18lookup}
\title{A model member lookup table for different UKCP18 collections}
\format{
A Dataframe object with 28 rows, 6 columns
}
\source{
\url{}
}
\usage{
ukcp18lookup
}
\description{
Matches member number with model name and in which collections / domains it is avaiolble
}
\keyword{datasets}
