% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukcp18sst}
\alias{ukcp18sst}
\title{A ~0.1 degree grid resolution dataset of UKCP18 sea-surface temperature data for NW Europe}
\format{
A PackedSpatRaster object with 375 rows, 297 columns and 3 layers
}
\source{
\url{ftp.ceda.ac.uk}
}
\usage{
ukcp18sst
}
\description{
A spatial dataset of monthly sea-surface temperatures for April to June 2018 for sea around
UK & North West Europe, covering the area bounded by  -19.9, 13.1, 40.0, 65.0
(xmin, xmax, ymin, ymax) with the WGS84 lat long coordinate reference system (EPSG:4326)
}
\keyword{datasets}
