% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{ukcp18toclimarray}
\alias{ukcp18toclimarray}
\title{convert UKCP18 ncdf4 files to format required for model}
\usage{
ukcp18toclimarray(
  dir_ukcp,
  dtm,
  startdate,
  enddate,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  wsalbedo = 0.19,
  bsalbedo = 0.22,
  ukcp_vars = c("clt", "hurs", "pr", "psl", "rls", "rss", "tasmax", "tasmin", "uas",
    "vas"),
  ukcp_units = c("\%", "\%", "mm/day", "hPa", "watt/m^2", "watt/m^2", "degC", "degC",
    "m/s", "m/s"),
  output_units = c("\%", "\%", "mm/day", "kPa", "watt/m^2", "watt/m^2", "degC", "degC",
    "m/s", "m/s"),
  temp_hgt = 1.5,
  wind_hgt = 10,
  toArrays = TRUE,
  sampleplot = FALSE
)
}
\arguments{
\item{dir_ukcp}{directory holding ALL ncdf files of UKCP18 data required to extract the requested variables and timeseries.}

\item{dtm}{SpatRaster object of elevations covering the geographical extent required (see details)}

\item{startdate}{POSIXlt class defining start date of required timeseries}

\item{enddate}{POSIXlt class defining end date of required timeseries}

\item{collection}{text string defining UKCP18 collection, either 'land-gcm' or 'land-rcm'}

\item{domain}{text string defining UKCP18 domain, either 'uk' or 'eur'(land-rcm collection only) or 'global'}

\item{member}{string defining the climate model member to be used for the timeseries. Available members vary between UKCP18 collections.}

\item{ukcp_vars}{UKCP18 variable names to be extracted DO NOT CHANGE}

\item{ukcp_units}{units of the UKCP18 variables extracted DO NOT CHANGE}

\item{output_units}{units required for output DO NOT CHANGE CHECK CORRECT}

\item{temp_hgt}{height of output temperature above ground (in metres)}

\item{wind_hgt}{height of output wind speed above ground (in metres)}

\item{toArrays}{logical determining if climate data returned as list of arrays. If FALSE returns list of Spatrasts.}

\item{sampleplot}{if TRUE plots examples of interpolated dates when converting from 360 to 366 day years}
}
\value{
a list of the following:
\describe{
\item{dtm}{Digital elevation of downscaled area in metres (as Spatraster)}
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{windheight_m}{Height of windspeed data in metres above ground (as numeric)}
\item{tempheight_m}{Height of temperature data in metres above ground (as numeric)}
\item{temp}{Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swrad}{Total downward shortwave radiation (W/m^2)}
\item{difrad}{Downward diffuse radiation (W / m^2)}
\item{lwrad}{Total downward longwave radiation (W/m^2)}
\item{windspeed}{At \code{windheight_m} above ground` (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
}
\description{
Converts UKCP18 global or regional land data in the form of netCDF4 file (as returned
by \code{\link[=download_ukcp18]{download_ukcp18()}} to the correct data form required for subsequent modelling.
}
\details{
The function converts 360 day UKCP18 to actual calendar data by reassignment and linear interpolation of missing data.
Returned climate data will be at the extent, corrdinate reference system of the UKCP18 data requested, which can vary between domains and collections.
The extent will be provided by the \code{dtm} which, if necessary, will be re-projected and resampled to the CRS and resolution of UKCP18 data requested.
For regional UKCP18 data, it is recommended that 'dtm' is derived from the original orography data available for download.
The dtm may also be used to estimate albedo on the basis of whether land or sea.
}
\examples{
 \dontrun{
dir_ukcp <- tempdir()
cedausr<-"your_user_name"
cedapwd <- "your_password"
startdate<-as.POSIXlt('2018-05-01')
enddate<- as.POSIXlt('2018-05-31')
collection<-'land-rcm'
domain<-'uk'
member<-'01'
rcp<-'rcp85'
vars<-c('clt','hurs','pr','psl','rls','rss','tasmax','tasmin','uas','vas')
download_ukcp18(dir_ukcp,startdate,enddate,collection,domain,'rcp85',member,vars,download_dtm=TRUE, cedausr,cedapwd)
dtm<-terra::rast(system.file('extdata/ukcp18rcm/orog_land-rcm_uk_12km_osgb.nc',package='mesoclim'))
dtm<-crop(dtm,c(28077.86, 339436.5, -57947.13, 204298.7))
ukcpinput<-ukcp18toclimarray(dir_ukcp,dtm,startdate, enddate, collection,domain,member)
}
}
\keyword{preprocess}
\keyword{ukcp18}
