% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukcpinput}
\alias{ukcpinput}
\title{A list of 12km resolution UKCP18 regional climate data}
\format{
a list with the following elements (climate variables as 3D arrays):
\describe{
\item{dtm}{a wrapped SpatRast of UKCP18 land elevations (m) matching the climate data; NA where UKCP18 classes a cell as sea}
\item{tme}{POSIXlt object of daily dates}
\item{windheight_m}{numeric value in metres of wind height above ground}
\item{tempheight_m}{numeric value in metres of temperature height above ground}
\item{cloud}{Cloud cover (Percentage)}
\item{relhum}{Relative humidity (Percentage)}
\item{prec}{Precipitation (mm/day)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{lwrad}{Downward longwave radiation (W/m^2)}
\item{swrad}{Downward shortwave radiation (W/m^2)}
\item{tmax}{Maximum daily temperature (deg C)}
\item{tmin}{Minimum daily temperature (deg C)}
\item{windspeed}{Wind speed (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
}
}
\source{
\url{https://catalogue.ceda.ac.uk}. Created by \code{data-raw/ukcp_sample_data.R}.
}
\usage{
ukcpinput
}
\description{
A list of daily UKCP18 regional climate model data (rcp85, member 01) for May 2018 for
Southwest UK, as returned by \code{\link[=ukcp18toclimarray]{ukcp18toclimarray()}} with collection='land-rcm' and domain='uk'.
Covers a grid of 4 x 3 cells of 12 km (x 144000-192000, y 0-36000, OSGB), which includes
the example DTMs \code{dtmf.tif} and \code{lizard50m.tif} in \code{inst/extdata/dtms}.
}
\keyword{datasets}
