% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{waterbodyeffects}
\alias{waterbodyeffects}
\title{Compute coastal effects}
\usage{
waterbodyeffects(
  tme,
  temperature,
  windspeed,
  winddir,
  wst,
  demf,
  demc = NA,
  method = 1,
  tasl = FALSE,
  steps = 8,
  use.raster = TRUE,
  zmin = 0,
  plot.progress = TRUE,
  tidyr = FALSE
)
}
\arguments{
\item{tme}{an POSIXlt object of times corresponding to values in \code{temperature}.}

\item{temperature}{a vector or array of hourly temperatures (deg C). Must be an array of \code{method > 1} (see details).}

\item{windspeed}{a vector or array of hourly wind speeds in m/s (see details).}

\item{winddir}{a vector or array of hourly wind directions (decimal degrees) in m/s (see details).}

\item{wst}{a vector, matrix or array of water body surface temperatures in deg C or NA (see details).}

\item{demf}{a RasterLayer or SpatRaster object of elevations (m) matching the spatial resolution for which outputs are needed}

\item{demc}{a RasterLayer or SpatRaster object of elevations (m) matching the spatial resoltuion of input temperature data. Ignored if method = 1.}

\item{method}{Optional integer in range 1-3 indicating which method is used to derive water body effects (see details)}

\item{tasl}{optional logical indicating whether values in \code{temperature} correpsond to sea-level or
include an altitudinal effects (see details).}

\item{steps}{an optional integer. Coastal effects are calculated in specified directions upwind. Steps defines the total number of directions used. If the default 8 is specified, coastal effects are calculated at 45º intervals.}

\item{use.raster}{an optional logical value indicating whether to mask the output values by \code{demf}.}

\item{zmin}{optional assumed sea-level height. Values below this are set to zmin}

\item{plot.progress}{logical value indicating whether to produce plots to track progress.}

\item{tidyr}{logical value indicating whether to download 30m resolution digital elevation data
and use these data to improve the assignment of land and sea pixels when calaculating coastal exposure}
}
\value{
a three dimensional array of temperatures with the first two dimensions matching x and y in \code{demf} and
the third dimension corresponding to times in \code{tme}
}
\description{
The function \code{waterbodyeffects} applies coastal effects using one of three user-specified
methods
}
\details{
if \code{wst = NA} water-bodies are assumed to be sea and sea-surface
temperature data are downloaded from the NOAA. If \code{method = 1}, Water body-land temperature differences are applied
using an empirical method calibrated against data obtained in Cornwall, UK.
If \code{method > 1} Thin-plate spline interpolation is used to derive coastal
effects with coastal exposure (upwind and mean in all directions)
as covariates). \code{Temperature} should be a three-dimensional array with x and y
dimensions matching those of \code{demc} Models are fitted with coarse-resolution
estimates of coastal exposure and applied at high-resolution using \code{demf}. To
ensure coastal effects capture waterbodies beyond the study area, digital
elevation data are automatically downloaded and any elevations < \code{zmin} are
assumed to correspond to waterbodies. Future versions of this function will
allow users to supply their own wide extend water-body data. The Thin-plate
spline methods are quite slow. If \code{method = 2} Thin-plate spline models are
fitted and applied at hours corresponding to daily maximum and minimum
temperatures only and the diurnal cycle in \code{temps} is applied to derive values
for hours in between. If \code{method = 3}, Thin-plate spline models are applied
hourly.

The Thin-plate spline models derive estimates of land-water-body
temperature differences. \code{wst} can be supplied as a vector of hourly values
(assumes no spatial variability), a matrix of values (assumes no temporal
variability) or as a 3D array of in which case the x and y dims must match
those of \code{temperature}. However if a 3D array is supplied there is no need for
the intervals to be hourly: hourly values are derived by spline interpolation.
As land-water body temperature differences are needed, if \code{wst} contains values
for water-body pixels only, values for land-areas are derived by interpolation.

If supplied as a 3D array the x and y dimensions of \code{temperature}, \code{windspeed} and
\code{winddir} and \code{sst} should match \code{demc}, but can also match \code{demf}. If \code{method > 1}
and \code{tasl = FALSE} elevational effects are removed by applying a fixed environmental
lapse rate of -5 degrees per 1000m prior to interpolation and the effect is
added back-in after downscaling.
}
