% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{write_climdata}
\alias{write_climdata}
\title{Write climate data}
\usage{
write_climdata(climdata, filepath, overwrite = FALSE)
}
\arguments{
\item{climdata}{a list of climate variables as output by \code{ukcp18toclimdata()}}

\item{filepath}{to be written}
}
\value{
writes an external file
}
\description{
Writes to a binary file a climate dataset (as output by for example \code{ukcp18toclimdata()}.
Wraps any Spatrasters to PackedSpatRasters before writing.
}
\examples{
climdata<-read_climdata(mesoclim::ukcpinput)
dir_temp<-tempdir()
write_climdata(climdata,file.path(dir_temp,'filename.rds'))
}
\keyword{data}
\keyword{preprocess}
