% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{write_parcels}
\alias{write_parcels}
\title{Write parcel climate .csv files}
\usage{
write_parcels(parcel_list, dir_out, overwrite = c("none", "append", "replace"))
}
\arguments{
\item{parcel_list}{\itemize{
\item list of parcel names and climate variables as output by \code{create_parcel_list}
}}

\item{dir_out}{\itemize{
\item directory to write to
}}

\item{overwrite}{\itemize{
\item string describing how to respond to existing files where 'none' stops if any file exists,
'overwrite' replaces existing files and 'append' adds results to existing files.
}}
}
\value{
NA - writes .csv files
}
\description{
Writes one file per parcel of climate timeseries with option to overwrite or append to existing files.
}
\keyword{data}
\keyword{postprocess}
