% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrect_functions.R
\name{biascorrect_climdata}
\alias{biascorrect_climdata}
\title{Apply bias correction models to all climate variables}
\usage{
biascorrect_climdata(
  climdata,
  model_list,
  prec_thold = 0.01,
  rangelims = NA,
  fillna = FALSE
)
}
\arguments{
\item{climdata}{\itemize{
\item a list of climate and associated data as output by \code{ukcp18toclimarray} function
}}

\item{model_list}{\itemize{
\item a list of bias correction models as output by \code{biascorrect} and \code{precipcorrect} functions
names of list elements must match variable names in \code{climdata}.
}}

\item{prec_thold}{\itemize{
\item numeric value (mm per day) below which daily precipitation is taken to be zero.
}}

\item{rangelim}{fractional amount above the maximum values of precipitation
in model observed data in any given cell by which values in the returned dataset can
attain. I.e. if set to 2, values in any given grid cell of the returned dataset
cannot exceed twice the maximum amount in the corresponding grid cell in
observed data of model.If not supplied (the default), returned values are unbounded.}
}
\value{
list of bias corected spatRasters and other variables matching \code{climdata} input
}
\description{
Apply bias correction models to all climate variables
}
\details{
SpatRaster output may be cropped if climdata input extent exceeds the extent over which bias correction
models were developed.
}
\keyword{biascorrect}
