% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{calculate_terrain_shading}
\alias{calculate_terrain_shading}
\title{Calculate horizon for different solar azimuths and total skyview}
\usage{
calculate_terrain_shading(dtm, steps = 24, toArrays = FALSE)
}
\arguments{
\item{toArrays}{}
}
\value{
list (length=2) of \code{skyview} and \code{horizon} arrays or SpatRasters
}
\description{
Calculate horizon for different solar azimuths and total skyview
}
\details{
Skyview places equal importance on each sector of sky
}
\examples{
results<-calculate_terrain_shading(dtmf)
plot(results$skyview)
plot(results$horizon[[c(1,6,12,18)]])
}
