% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{checkinputs}
\alias{checkinputs}
\title{Check and summarise coarse resolution climate data inputs}
\usage{
checkinputs(input_list, tstep = c("hour", "day"), plots = TRUE)
}
\arguments{
\item{input_list}{list of climate and associated variables as output by functions like 'ukcp18toclimarray'}

\item{tstep}{string decribing whether input data is by hour or day}
}
\value{
where possible returns inputs suitable for downsclaing and/or warning messages if input data has unexpected values
Also prints summary statistics and graphs of each variable
}
\description{
Check and summarise coarse resolution climate data inputs
}
\examples{
climdata<- read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
chk_climdata<- checkinputs(climdata, tstep = "day")
}
\keyword{preprocess}
