% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{create_ukcpsst_data}
\alias{create_ukcpsst_data}
\title{Create UKCP sea surface temperature rast stack from ncdf files.
Resamples and reprojects sea surface data to aoi as well as interpolating any missing coastal values.
Interpolation method = mean of all neighbouring cells.}
\usage{
create_ukcpsst_data(
  dir_data,
  startdate,
  enddate,
  dtmc = NA,
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15"),
  v = "SST"
)
}
\arguments{
\item{dir_data}{directory holding SST .nc files downloaded from ceda}

\item{startdate}{start date as POSIXlt}

\item{enddate}{end date as POSIXlt}

\item{v}{= SST sea surface temperature variable name}

\item{aoi}{SpatRaster, SpatVector or sf object defining area of interest to crop data. NA no cropping occurs}

\item{members}{model members to be included}
}
\value{
Spatraster timeseris of sea surface temperatures resampled to aoi projection and resolution
}
\description{
Create UKCP sea surface temperature rast stack from ncdf files.
Resamples and reprojects sea surface data to aoi as well as interpolating any missing coastal values.
Interpolation method = mean of all neighbouring cells.
}
\examples{
 \dontrun{
 dir_sst<-system.file("extdata/sst", package = "mesoclim")
 sst<-create_ukcpsst_data(dir_sst,as.POSIXlt('2018/05/01'),as.POSIXlt('2018/05/31'),dtmc=ukcpinput$dtm, member="01")
 plot(c(sst,ukcpinput$dtm))
 }
}
\keyword{preprocess}
\keyword{ukcp18}
