% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{create_ukcpsst_data}
\alias{create_ukcpsst_data}
\title{Create UKCP sea surface temperature rast stack from ncdf files}
\usage{
create_ukcpsst_data(
  dir_data,
  startdate,
  enddate,
  aoi = NA,
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15"),
  v = "SST"
)
}
\arguments{
\item{dir_data}{directory holding SST .nc files downloaded from ceda}

\item{startdate}{start date as POSIXlt}

\item{enddate}{end date as POSIXlt}

\item{aoi}{SpatRaster, SpatVector or sf object defining area of interest to crop data. NA no cropping occurs}

\item{v}{= SST sea surface temperature}

\item{members}{model members to be included}
}
\value{
Spatraster timeseris of sea surface temperatures
}
\description{
Create UKCP sea surface temperature rast stack from ncdf files
}
\examples{
dir_data<-system.file('extdata/ukcp18sst',package='mesoclim')
sst<-create_ukcpsst_data(dir_data,as.POSIXlt('2018/05/01'),as.POSIXlt('2018/05/31'),member='01')
}
\keyword{preprocess}
\keyword{ukcp18}
