% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.basindelin}
\alias{.basindelin}
\title{Internal function for delineating basins with option for boundary > 0}
\usage{
.basindelin(dtm, boundary = 0)
}
\description{
Internal function for delineating basins with option for boundary > 0
}
