% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.cad_conditions}
\alias{.cad_conditions}
\title{Calculate whether conditions allow for cold air drainage}
\usage{
.cad_conditions(tc, u2, swrad, lwrad, dtmc, dtmf, refhgt = 2)
}
\arguments{
\item{tc}{\itemize{
\item temperature in deg celsius
}}

\item{u2}{\itemize{
\item windspeed in m/s
}}

\item{lwrad}{\itemize{
\item LW downward radiation
}}

\item{dtmc}{\itemize{
\item coarse resolution dtm matching climate variable inputs - not required if other parameters all spatRasters.
}}

\item{dtmf}{\itemize{
\item fine scale dtm for outputs
}}

\item{refhgt}{\itemize{
\item of temp above ground (m)
}}

\item{swad}{\itemize{
\item SW downward radiation
}}
}
\value{
a binary spatRaster matching dtmf where 1= suitable, 0 = not suitable for cold air drainage
}
\description{
Calculate whether conditions allow for cold air drainage
}
\details{
temperature, wind  and radiation data used to determine
whether cold air drainage conditions are likely to occur and returns a binary
vector of the same length as \code{em} and \code{wind} indicating whether conditions
occur (1) or not (0). For daily measurements it may be appropriate to apply only to daily min
and  to set swrad to 0 assuming this minimum occurs during the night.
}
\keyword{internal}
