% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.resample}
\alias{.resample}
\title{version of terra resample that equates to NA.RM = TRUE}
\usage{
.resample(r1, r2, msk = FALSE, method = "bilinear")
}
\arguments{
\item{r1}{\itemize{
\item is resampled to same geometry as r2
}}

\item{r2}{\itemize{
\item target geometry
}}

\item{method}{for resample and project can be set}

\item{msk=TRUE}{if output to be masked out where r2 cells = NA}
}
\description{
version of terra resample that equates to NA.RM = TRUE
}
