% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.sea_to_coast}
\alias{.sea_to_coast}
\title{Interpolates sea surface temperature data to coastal regions
Resamples sst data to same extent, resolution and projection as aoi
Interpolates from mean of adjacent sea cells}
\usage{
.sea_to_coast(sst.r, aoi.r, ext_cells = 8)
}
\arguments{
\item{sst.r}{\itemize{
\item sea surface temperature data
}}

\item{aoi.r}{\itemize{
\item dtm of area of interest where sea = NA
}}
}
\value{
sst spatrasters of same projection, res and extent as aoi with matching coastline
}
\description{
Interpolates sea surface temperature data to coastal regions
Resamples sst data to same extent, resolution and projection as aoi
Interpolates from mean of adjacent sea cells
}
\examples{
sst.r<-rast(system.file("extdata/sst/NWSClim_NWSPPE_r001i1p00000_2018_gridT.nc",package="mesoclim"),"SST")[[1]]
aoi.r<-unwrap(mesoclim::ukcpinput$dtm)
plot(.sea_to_coast(sst.r,aoi.r))
}
