% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tempcad}
\alias{.tempcad}
\title{DEFUNCT Downscale temperature with cold air drainage effects}
\usage{
.tempcad(climdata, tempvar, lrf, dtmf, basins = NA, refhgt = 2)
}
\description{
DEFUNCT Downscale temperature with cold air drainage effects
}
\keyword{internal}
