% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tempcoastal}
\alias{.tempcoastal}
\title{Coastal temperature effects}
\usage{
.tempcoastal(tc, sstf, u2, wdir, dtmf, dtmm, dtmc)
}
\arguments{
\item{tc}{\itemize{
\item downscaled temperature at same res as dtmf
}}

\item{sstf}{\itemize{
\item downscaled sea surface temperature to dtmf resolution and extent - no NA and timeseries matches tc
}}

\item{u2}{\itemize{
\item downscaled windspeed at 2m height
}}

\item{wdir}{\itemize{
\item wind direction (coarse resolution) - same value for all of dtmf extent will be used
}}

\item{dtmf}{\itemize{
\item fine dtm spatraster
}}

\item{dtmm}{\itemize{
\item medium dtm spatraster
}}

\item{dtmc}{\itemize{
\item coarse dtm spatraster
}}
}
\value{
Spatraster of temperature that includes coastal effect
}
\description{
Coastal temperature effects
}
\keyword{internal}
