% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.windhgt}
\alias{.windhgt}
\title{Applies height correction to wind speed measurements}
\usage{
.windhgt(wspeed, zi, zo)
}
\arguments{
\item{zi}{a numeric value idicating the height (m) above the ground of \code{wspeed} input}

\item{zo}{a numeric value indicating the height (m) above ground level of output speeds
Assumes a logarithmic height profile and imposes a minimum zo of 0.2.
Equivalent of using a roughness length (z0) of ~0.2 where v = vref ln(z/z0)/ln(zref/z0) (https://www.rensmart.com/Information/WindSheer)}
}
\description{
cf microclima::windheight
wspeed numeric value(s) of measured wind speed (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{ m s^{-1}}}) at height \code{zi} (m).
}
\keyword{internal}
