% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_hadukdaily}
\alias{download_hadukdaily}
\title{Downloads 1km gridded historic climata data for the UK}
\usage{
download_hadukdaily(
  dir_out,
  startdate,
  enddate,
  varn = c("rainfall", "tasmax", "tasmin", "hurs", "psl", "pv", "sun", "sfsWind"),
  freq = c("day", "mon"),
  cedausr,
  cedapwd,
  cedaprot = "https://",
  cedaserv = "dap.ceda.ac.uk"
)
}
\arguments{
\item{dir_out}{directory to which to save data}

\item{startdate}{\itemize{
\item POSIXlt value indicating start date for data required
}}

\item{enddate}{\itemize{
\item POSIXlt value indicating end date for data required
}}

\item{varn}{vector of variables required, one or more of: 'rainfall','tasmax','tasmin','hurs','psl','pv','sun','sfsWind'.
NOTE: Only \code{rainfall} (precipitation in mm), \code{tasmax}(maximum daily temperature, deg C) and
\code{tasmin} (minimum daily temperature, deg C) are available at daily time steps (all other monthly only).}

\item{freq}{\itemize{
\item string of frequency required - either 'day' or 'mon' (monthly)
}}

\item{cedausr}{ceda username string}

\item{cedapwd}{ceda password sting}

\item{cedaprot}{\itemize{
\item ceda protocol used is https  - DO NOT CHANGE
}}

\item{cedaserv}{\itemize{
\item ceda service is dap - DO NOT CHANGE
}}
}
\value{
database detailing download details and success of all requested files
}
\description{
The function \code{download_hadukdaily} downloads HadUK-Grid Gridded
Climate Observation data on a 1km grid over the UK from CEDA.
}
\details{
Uses ceda dap https service that does not require username or password.
}
\examples{
\dontrun{dir_out <- tempdir()
cedausr<-"your_user_name"
cedapwd <- "your_password"
download_hadukdaily(dir_out,as.POSIXlt('2018-05-01'),as.POSIXlt('2018-05-31'),varn=c('tasmax','tasmin'),freq='day', cedausr,cedapwd)
}
}
\keyword{download}
