% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_ukcp18}
\alias{download_ukcp18}
\title{Download UKCP18 climate data}
\usage{
download_ukcp18(
  dir_out,
  cedausr,
  cedapwd,
  startdate,
  enddate,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  rcp = c("rcp85", "rcp26"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  vars = c("clt", "hurs", "huss", "pr", "prsn", "psl", "rls", "rss", "snw", "tasmax",
    "tasmin", "uas", "vas"),
  cedaprot = "ftp://",
  cedaserv = "ftp.ceda.ac.uk"
)
}
\arguments{
\item{dir_out}{Output directory to which files are downloaded}

\item{cedausr}{string of ceda user name}

\item{cedapwd}{string of ceda user pasword}

\item{startdate}{POSIXlt class defining start date of required timeseries}

\item{enddate}{POSIXlt class defining end date of required timeseries}

\item{collection}{text string defining UKCP18 collection, either 'land-gcm' or 'land-rcm'}

\item{domain}{text string defining UKCP18 domain, either 'uk' or 'eur'(land-rcm collection only) or 'global'}

\item{rcp}{text string iof RCP scenario to be downloaded}

\item{member}{vector of strings defining the climate model member to be downloaded. Available members vary between UKCP18 collections.}

\item{vars}{vector of strings corresponding to UKCP18 short variable names to download}

\item{cedaprot}{string of ceda protocol set to ftp DO NOT CHANGE}

\item{cedaserv}{string of the ceda server url DO NOT CHANGE}
}
\value{
downloads files to \code{dir_out}
}
\description{
Using parameters the function will downloaded from ftp.ceda.ac.uk all available UCKP18 files containing relevant data to the user defined output directory
}
