% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{download_ukcpsst}
\alias{download_ukcpsst}
\title{Download UKCP18rcm driven sea surface temperature predictions}
\usage{
download_ukcpsst(
  dir_out,
  startdate,
  enddate,
  modelruns = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  cedausr,
  cedapwd,
  access_token
)
}
\arguments{
\item{dir_out}{\itemize{
\item local directory to which files written
}}

\item{startdate}{\itemize{
\item start date as POSIXlt
}}

\item{enddate}{\itemize{
\item end date as POSIXlt
}}

\item{modelruns}{\itemize{
\item vector of two character strings of model numbers
}}

\item{cedausr}{ceda username string}

\item{cedapwd}{ceda password sting}

\item{access_token}{ceda access token as provided by \href{https://services-beta.ceda.ac.uk/account/token/?_ga=2.221117848.1874942177.1727205775-405001952.1710779653}{Ceda access token generator}}
}
\value{
downloaded file paths
}
\description{
For description of the modelling used to produce sea surface temperatures then see:
\verb{https://www.metoffice.gov.uk/binaries/content/assets/metofficegovuk/pdf/research/ukcp/ukcp18-guidance-data-availability-access-and-formats.pdf}
}
\examples{
\dontrun{
 cedausr<-"your_user_name"
 cedapwd <- "your_password"
 access_token<-"your_access_token"
 startdate<-as.POSIXlt('2017/12/31')
 enddate<-as.POSIXlt('2018/12/31')
 modelruns<-c('01')
 dir_out<-tempdir()
 download_ukcpsst(dir_out,startdate,enddate,modelruns, cedausr,cedapwd,access_token)
 }
}
\keyword{download}
\keyword{ukcp18}
