% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{era5toclimarray}
\alias{era5toclimarray}
\title{convert era4 ncdf4 file to format required for model}
\usage{
era5toclimarray(ncfile, dtm, dtr_cor_fac = 1.285)
}
\arguments{
\item{ncfile}{character vector containing the path and filename of the nc file}

\item{dtm}{a SpatRaster object of elevations covering the extent of the study area (see details)}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction of coastal grid cells. Default = 1.285, based on calibration against UK Met Office
observations. If set to zero, no correction is applied.}
}
\value{
a list of the following:
\describe{
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{climarray}{a list of arrays of hourly weather variables
\item{dtmc}{a coarse resolution digital elevation dataset matching the resolution of input
climate data, but with a coordinate reference system and extent matching \code{dtm}}
}
}
\description{
The function \code{era5toclimarray} converts data in a netCDF4 file returned
by \code{\link[mcera5:request_era5]{mcera5::request_era5()}} to the correct formal required for subsequent modelling.
}
\details{
the model requires that input climate data are projected using a coordinate reference
system in which x and y are in metres. Since values returned by \code{\link[mcera5:request_era5]{mcera5::request_era5()}}
are in lat long, the output data are reprojected using the coordinate reference system and
extent of dtm (but retain the approximate original grid resolution of the input climate data).
Returned climate data match the resolution, corrdinate reference system and extent of \code{dtmc}.
}
