% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_functions.R
\name{era5toclimarray}
\alias{era5toclimarray}
\title{convert era4 ncdf4 file to format required for model}
\usage{
era5toclimarray(
  ncfile,
  dtm = NA,
  aoi = NA,
  dtr_cor_fac = 1.285,
  toArrays = TRUE
)
}
\arguments{
\item{ncfile}{character vector containing the path and filename of the nc file}

\item{dtm}{a SpatRaster object of elevations covering the extent of the study area (see details)}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction of coastal grid cells. Default = 1.285, based on calibration against UK Met Office
observations. If set to zero, no correction is applied.}

\item{toArrays}{logical determining if climate data returned as list of arrays. If FALSE returns list of Spatrasters.}
}
\value{
a list of the following:
\describe{
\item{dtm}{Digital elevation of downscaled area in metres (as Spatraster)}
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{windheight_m}{Height of windspeed data in metres above ground (as numeric)}
\item{tempheight_m}{Height of temperature data in metres above ground (as numeric)}
\item{temp}{Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swrad}{Total downward shortwave radiation (W/m^2)}
\item{difrad}{Downward diffuse radiation (W / m^2)}
\item{lwrad}{Total downward longwave radiation (W/m^2)}
\item{windspeed}{at 2m above ground (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
}
\description{
The function \code{era5toclimarray} converts data in a netCDF4 file returned
by \code{\link[mcera5:request_era5]{mcera5::request_era5()}} to the correct formal required for subsequent modelling.
}
\details{
the model requires that input climate data are projected using a coordinate reference
system in which x and y are in metres. Since values returned by \code{\link[mcera5:request_era5]{mcera5::request_era5()}}
are in lat long, the output data are reprojected using the coordinate reference system and
extent of dtm (but retain the approximate original grid resolution of the input climate data).
Returned climate data match the resolution, coordinate reference system and extent of \code{dtmc}.
}
\examples{
ncfile<-system.file('extdata/era5/era5_surface_ukeire_2018_05.nc',package='mesoclim')
aoi<-terra::vect(terra::ext(-7.125,-2.875,49.375,51.625),crs='EPSG:4326')
dtm<-terra::rast(system.file('extdata/dtms/era5dtm.tif',package='mesoclim'))
era5input<-era5toclimarray(ncfile, dtm=NA, aoi=aoi)
plot_q_layers(.rast(era5input$temp,era5input$dtm))
checkinputs(era5input,'hour')
}
\keyword{era5}
\keyword{preprocess}
