% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{lapserate}
\alias{lapserate}
\title{Calculate Lapse rates}
\usage{
lapserate(tc, ea, pk)
}
\arguments{
\item{ea}{= vapour pressure (kPa)}

\item{pk}{= atmospheric pressure (kPa)}
}
\value{
laps rate
}
\description{
Calculate Lapse rates
}
\examples{
climdata<-read_climdata(system.file('data/era5input.rds',package='mesoclim'))
ea<-converthumidity(climdata$relhum,tc=climdata$temp , pk=climdata$pres)
lr<-lapserate(climdata$temp,ea,climdata$pres)
plot(.rast(lr,climdata$dtm)[[1]])
}
