% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{lapserate}
\alias{lapserate}
\title{Calculate Lapse rates
ea = temperature (deg C)
ea = vapour pressure (kPa)
pk = atmospheric pressure (kPa)}
\usage{
lapserate(tc, ea, pk)
}
\description{
Calculate Lapse rates
ea = temperature (deg C)
ea = vapour pressure (kPa)
pk = atmospheric pressure (kPa)
}
