% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{lw_dailytohourly}
\alias{lw_dailytohourly}
\title{Daily to hourly downward longwave radiation}
\usage{
lw_dailytohourly(
  lw = daily100m$lwrad,
  dtm = daily100m$dtm,
  hrtemps = hrtemps,
  hrrh = hrrh,
  hrpres = hrpres,
  tme,
  adjust = FALSE
)
}
\arguments{
\item{lw}{\itemize{
\item an array of daily mean sky-emissivity values values (0-1)
}}

\item{dtm}{\itemize{
\item digital elevation spatraster
}}

\item{hrtemps}{\itemize{
\item hourly temperature data for corresponding period as lw
}}

\item{hrrh}{\itemize{
\item hourly relative humidity data for corresponding period as lw
}}

\item{hrpres}{\itemize{
\item hourly atmospheric pressure data for corresponding period as lw
}}

\item{tme}{\itemize{
\item POSIXlt object of dates corresponding to radsw
}}

\item{adjust}{\itemize{
\item optional logical which if TRUE ensures that, after interpolation, returned
hourly values, when averaged to daily, match the input
}}
}
\value{
an array of hourly downward long wave radiation
}
\description{
Derives an array or spatraster of hourly effective downward longwave radiation values.
}
\details{
Effective sky emissvity can be used to calaculate downward longwave radiation (Lwd).
The formula is Lwd = skyem * Lwu where Lwu is upward longwave radiation given
by Lwu=0.97\emph{5.67}10**-8*(tc+273.15). Here tc is average surface temperature (deg C))
but an adequate approximation is derived if subtited by air temperature.
TO DO - Option of providing cloud cover data to correct sky emissivity values??
}
\examples{
lwhr<- lw_dailytohourly(lw=daily100m$lwrad, dtm=daily100m$dtm, hrtemps=hrtemps, hrrh=hrrh, hrpres=hrpres, tme=daily100m$tme,  adjust = FALSE)
}
\keyword{temporal}
