% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{lwdownscale}
\alias{lwdownscale}
\title{Downscale longwave radiation}
\usage{
lwdownscale(lwrad, tc, tcf, tme, dtmf, dtmc, terrainshade = FALSE)
}
\arguments{
\item{lwrad}{\itemize{
\item coarse resolution downward longwave radiation (W/m^2)  as spatraster or array.
}}

\item{tc}{\itemize{
\item coarse resolution temperature (deg C) as spatraster or array.
}}

\item{tcf}{\itemize{
\item fine resolution spatraster of temperature (deg C) down-scaled to resolution of \code{dtmf}.
}}

\item{tme}{POSIXlt object of times corresponding to radiation values in
\code{lwrad}, \code{tc} and \code{tcf}.}

\item{dtmf}{a fine-resolution SpatRast of elevations.}

\item{dtmc}{SpatRast of elevations matching the resolution, extend and coordinate
reference system of \code{lwrad} and \code{tc}.}

\item{terrainshade}{\itemize{
\item boolean TRUE/FALSE
}}
}
\value{
a multi-layer SpatRast of downward longwave radiation (W/m^2)
matching the resolution of dtmf.
}
\description{
Downscales an array of longwave downward radiation data with option to
simulate terrain shading of skyview
}
\details{
radiation is downscaled by computing from temperature the difference in upward
longwave radiation at fine and coarse resolutions. Corrected LW down is then
corrected for terrain shading of skyview if requested.
}
