% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{prec_dailytohourly}
\alias{prec_dailytohourly}
\title{Daily to hourly precipitation downscale (TEMPORARY HOLDING FUNCTION)}
\usage{
prec_dailytohourly(prec, tme = NA, dailyraincut = 0)
}
\arguments{
\item{tme}{}
}
\description{
Daily to hourly precipitation downscale (TEMPORARY HOLDING FUNCTION)
}
\examples{
climdaily<-read_climdata(ukcpinput)
hrprec<-prec_dailytohourly(climdaily$prec)
#Plot results for one cell
cell_prec<-t(terra::extract(hrprec,matrix(c(175000,40000),ncol=2)))
matplot(terra::time(hrprec),cell_prec, type = "l", lty = 1)
}
