% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporaldownscale.R
\name{pres_dailytohourly}
\alias{pres_dailytohourly}
\title{Daily to hourly atmospheric pressure}
\usage{
pres_dailytohourly(pres, tme, adjust = TRUE)
}
\arguments{
\item{pres}{\itemize{
\item an array of daily mean surface-level pressure values (kPa)
}}

\item{tme}{\itemize{
\item POSIXlt object of dates corresponding to radsw
}}

\item{adjust}{\itemize{
\item optional logical which if TRUE ensures that, after interpolation, returned
hourly values, when averaged to daily, match the input
}}
}
\value{
an array of hourly surface-level pressure values (kPa)
}
\description{
Derives an array of hourly surface-level pressure values from an array of daily
surface level pressure values.
}
\details{
Note that, as in the era5 dataste, these are surface level pressures in kPa corrsponding to
the elevation of each geographic location and not sea=level pressures. The
conversion to millbars is mb = 10 kPa. The conversion to sea-level pressure (Psl) can be
derived by reversing the equation Psl = pres*((293-0.0065z)/293)**5.26
}
\examples{
climdata<- read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
presh<-pres_dailytohourly(climdata$pres,climdata$tme)
plot_q_layers(.rast(presh,climdata$dtm))
}
\keyword{temporal}
