% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{presdownscale}
\alias{presdownscale}
\title{Downscale pressure with elevation effects}
\usage{
presdownscale(pk, dtmf, dtmc, sealevel = TRUE)
}
\arguments{
\item{pk}{a SpatRast or array of atmospheric pressures (kPa). If an array dtmc must
be provided.}

\item{dtmf}{a fine-resolution SpatRast of elevations. Temperatures down-scaled
to resolution of \code{dtmf}.}

\item{dtmc}{optional SpatRast of elevations matching resolution of \code{pk}. If
not supplied, \code{pk} must be a SpatRast and \code{dtmc} is derived by resampling \code{dtmf}
to resolution of \code{pk} and extents must match.}

\item{sealevel}{optional logical indicating whether \code{pk} is pressure at sea-level}
}
\value{
a multi-layer SpatRast of elevation-corrected pressures (kPa) matching the
resolution of dtmf.
}
\description{
Downscales an array of pressure data applying elevation effects
}
\examples{
climdata<- read_climdata(mesoclim::ukcpinput)
pk <- presdownscale(climdata$pres, terra::rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim')), climdata$dtm)
terra::panel(c(pk[[1]],pk[[5]],pk[[10]],pk[[20]]),main=paste0("Atmospheric pressure ",c(1,5,10,15),"/05/2018"))
}
\keyword{spatial}
