% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatefunctions.R
\name{sea_to_atmos_pressure}
\alias{sea_to_atmos_pressure}
\title{Convert sea to atmospheric pressure}
\usage{
sea_to_atmos_pressure(psl, dtm)
}
\arguments{
\item{psl}{\itemize{
\item numeric or spatraster sea level pressure
}}

\item{dtm}{\itemize{
\item numeric or spatraster elevation
}}
}
\value{
numeric or spatraster of sea level pressures matching dtm
}
\description{
Convert sea to atmospheric pressure
}
