% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{spatialdownscale}
\alias{spatialdownscale}
\title{Spatially downscale all climate variables}
\usage{
spatialdownscale(
  climdata,
  SST,
  dtmf,
  dtmm = NA,
  basins = NA,
  cad = TRUE,
  coastal = TRUE,
  refhgt = 2,
  uhgt = 2,
  rhmin = 20,
  pksealevel = TRUE,
  patchsim = TRUE,
  terrainshade = TRUE,
  precipmethod = "Elev",
  fast = TRUE,
  noraincut = 0
)
}
\arguments{
\item{climdata}{a \code{climdata} model object containing climate data of the same format as \code{era5climdata}}

\item{SST}{a coarse resolution SpatRast of hourly sea-surface temperature data (deg C) with no NAs, as
returned by \code{\link[=SSTinterpolate]{SSTinterpolate()}}}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf (only needed for coastal effects - see details under \code{\link[=tempdownscale]{tempdownscale()}}).}

\item{basins}{optionally, a fine-resolution SpatRast of basins as returned by \code{\link[=basindelin]{basindelin()}}
matching the coordinate reference system and extent of \code{dtmf}. Calculated if
not supplied.}

\item{cad}{optional logical indicating whether to calculate cold-air drainage effects}

\item{coastal}{optional logical indicating whether to calculate coastal effects}

\item{refhgt}{height above ground of temperature measurements in \code{climdata}.}

\item{uhgt}{height above gorund of wind speed measurements in \code{climdata}
to resolution of \code{dtmf}.}

\item{rhmin}{minimum relative humidity (set to avoid relative humidity dropping too low
during down-scaling). Default 20 (percent).}

\item{pksealevel}{optional logical indicating whether input pressure data represent
sea-level pressure (default TRUE).}

\item{terrainshade}{optional logical indicating whether to account for terrain shading
during shortwave radiation downscale.}

\item{precipmethod}{One of \code{Tps} or \code{Elev} indicating whether to account for
elevation effects using a Thin-plate spline model or en emprical adjustment
(see details under \code{\link[=precipdownscale]{precipdownscale()}})}

\item{fast}{optional logical indicating whether to use fast Thin-plate spline
down-scaling (see details under \code{\link[=precipdownscale]{precipdownscale()}}).}

\item{noraincut}{numeric value below which low precipitation amounts are set to
zero (see details under \code{\link[=precipdownscale]{precipdownscale()}}).}

\item{pathsim}{optional logical indicating whether to simulate cloud cover and
rainfall patchiness see details under \code{\link[=swdownscale]{swdownscale()}} and \code{\link[=precipdownscale]{precipdownscale()}}.}

\item{difr}{optional logical indicating whether to return diffuse radiation (
default TRUE).}
}
\value{
a list of multi-layer SpatRast of downscaled climate variables as follows:
\describe{
\item{temp}{Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swrad}{Total downward shortwave radiation (W/m^2)}
\item{difrad}{Downward diffuse radiation (W / m^2)}
\item{lwrad}{Total downward longwave radiation (W/m^2)}
\item{windspeed at 2m (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
If input data are hourly \code{temp} is replaced with \code{tmax} and \code{tmin} representing
daily maximum and minimum values. The variable \code{difrad} is not returned unless
\code{terrainshade = TRUE}.
}
\description{
Spatially downscales coarse-resolution climate data
}
