% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{subset_climdata}
\alias{subset_climdata}
\title{Subset climate data by date}
\usage{
subset_climdata(climdata, sdatetime, edatetime)
}
\arguments{
\item{climdata}{\itemize{
\item list of spatraster/arrays and associated climate data
}}

\item{sdatetime}{\itemize{
\item start date time as POSIXlt
}}

\item{edatetime}{\itemize{
\item end date time as POSIXlt
}}
}
\description{
Subset climate data by date
}
\keyword{preprocess}
