% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{swdownscale}
\alias{swdownscale}
\title{Downscale shortwave radiation}
\usage{
swdownscale(swrad, tme, dtmf, dtmc, patchiness = 0)
}
\arguments{
\item{swrad}{a SpatRast or array of shortwave radiation (W/m^2). If an array dtmc must
be provided.}

\item{tme}{POSIXlt object of times corresponding to radiation values in \code{swrad}.}

\item{dtmf}{a fine-resolution SpatRast of elevations. Temperatures down-scaled
to resolution of \code{dtmf}.}

\item{dtmc}{optional SpatRast of elevations matching resolution of \code{swrad}. If
not supplied, \code{swrad} must be a SpatRast and \code{dtmc} is derived by resampling \code{dtmf}
to resolution of \code{swrad} and extents must match.}

\item{patchiness}{one of 0 (not simulated), 1 (low patchiness), 2 (medium patchines)
or (3) high patchiness}
}
\value{
a multi-layer SpatRast of shortwave radiation (W/m^2) matching the
resolution of dtmf.
}
\description{
Downscales an array of shortwave radiation data with option to
simulate cloud patchiness
}
