% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{tempdaily_downscale}
\alias{tempdaily_downscale}
\title{Downscale daily temperatures}
\usage{
tempdaily_downscale(
  climdata,
  tmean = NA,
  sst = NA,
  dtmf,
  dtmm,
  basins,
  uzf,
  cad = TRUE,
  coastal = TRUE,
  thgto = 2,
  whgto = 2
)
}
\arguments{
\item{climdata}{a \code{climdata} model object containing climate data of the same format as \code{era5climdata}
Expects variables to include a \code{tmin} and \code{tmax}}

\item{tmean}{a coarse resolution array or spatRaster of daily mean temperatures identical to climdata inputs.
If NA will calculate from hourluy downscaling of tmin and tmax in \code{climdata}}

\item{sst}{a vector or SpatRast of sea-surface temperature data (deg C) timeseries of same length or layer number as timesteps in climdata}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf (only needed for coastal effects - see details).}

\item{basins}{optionally, a fine-resolution SpatRast of basins as returned by \code{\link[=basindelin]{basindelin()}}
matching the coordinate reference system and extent of \code{dtmf}. Calculated if
not supplied.}

\item{uzf}{optionally, a SpatRast of fine resolution wind speeds as returned by \code{\link[=winddownscale]{winddownscale()}}.
Calculated if not supplied.}

\item{cad}{optional logical indicating whether to calculate cold-air drainage effects}

\item{coastal}{optional logical indicating whether to calculate coastal effects}

\item{thgto}{height above ground of output temperature measurements.}

\item{whgto}{height above ground of output wind speed measurements.
to resolution of \code{dtmf}.}

\item{tempvar}{string name of element of climdata holding temperature data (ie 'temp','tmax','tmin')}
}
\value{
a list of three multi-layer SpatRasters of downscaled daily minimum, maximum and mean temperatures (deg C) matching the
resolution of dtmf.
}
\description{
Downscales daily coarse-resolution temperature stats accounting for elevation
effects and optionally cold air drainage and coastal effects
}
\details{
Cold air drainage is calculated by delineating hydrological basins and
calculating flow accumulation and the elevation difference from the highest point of the basin.
Cold-air drainage is assumed to occur when atmospheric stability is high, namely when
the radiation balance is negative and wind speeds are low. Coastal effects are
calculated by determining the ratio of land to sea pixels in an upwind direction.
The provision of \code{dtmm} allows this ratio to be derived accoutning for land and
sea outside the boundaries of the study area.
}
\examples{
climdata<-read_climdata(mesoclim::ukcpinput)
dtmf<-terra::rast(system.file("extdata/dtms/dtmf.tif",package="mesoclim"))
dtmm<-terra::rast(system.file("extdata/dtms/dtmm.tif",package="mesoclim"))
basins<-basindelin(dtmf, boundary = 2)
wsf<-winddownscale(climdata$windspeed, climdata$winddir, dtmf, dtmm, climdata$dtm, zi=climdata$windheight_m)
dailytemps<-tempdaily_downscale(climdata,NA,terra::unwrap(mesoclim::ukcp18sst),dtmf,dtmm,basins,wsf,cad = TRUE,coastal = TRUE,2,2)
terra::panel(c(dailytemps$tmin[[13]],dailytemps$tmax[[13]],dailytemps$tmean[[13]]),main=paste(c("Tmin","Tmax","Tmean"),"13/05/2018"))
# With sea temperature as a vector
sst<-seq(10,13.1,length=31)
dailytemps<-tempdaily_downscale(climdata,NA,sst,dtmf,dtmm,basins,wsf,cad = TRUE,coastal = TRUE,2,2)
}
\keyword{spatial}
