% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{tempdownscale}
\alias{tempdownscale}
\title{Downscale temperature}
\usage{
tempdownscale(
  climdata,
  sst,
  dtmf,
  dtmm = NA,
  basins = NA,
  u2 = NA,
  cad = TRUE,
  coastal = TRUE,
  tempvar = "temp",
  thgto = 2,
  whgto = 2
)
}
\arguments{
\item{climdata}{a \code{climdata} model object containing climate data of the same format as \code{era5climdata}}

\item{sst}{a SpatRast of sea-surface temperature data (deg C) timeseries that overlaps climdata$tme}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf (only needed for coastal effects - see details).}

\item{basins}{optionally, a fine-resolution SpatRast of basins as returned by \code{\link[=basindelin]{basindelin()}}
matching the coordinate reference system and extent of \code{dtmf}. Calculated if
not supplied.}

\item{u2}{optionally, a SpatRast of high resolution wind speeds as returned by \code{\link[=windownscale]{windownscale()}}.
Calculated if not supplied.}

\item{cad}{optional logical indicating whether to calculate cold-air drainage effects}

\item{coastal}{optional logical indicating whether to calculate coastal effects}

\item{tempvar}{string name of element of climdata holding temperature data (ie 'temp','tmax','tmin')}

\item{thgto}{height above ground of output temperature measurements.}

\item{whgto}{height above ground of output wind speed measurements.
to resolution of \code{dtmf}.}
}
\value{
a multi-layer SpatRast of downscaled temperatures (deg C) matching the
resolution of dtmf.
}
\description{
Downscales coarse-resolution temperature data accounting for elevation
effects and optionally cold air drainage and coastal effects
}
\details{
Cold air drainage is calculated by delineating hydrological basins and
calculating flow accumulation and the elevation difference from the highest point of the basin.
Cold-air drainage is assumed to occur when atmospheric stability is high, namely when
the radiation balance is negative and wind speeds are low. Coastal effects are
calculated by determining the ratio of land to sea pixels in an upwind direction.
The provision of \code{dtmm} allows this ratio to be derived accoutning for land and
sea outside the boundaries of the study area.
}
\examples{
# Takes ~90 seconds to run
 dir_data<-system.file('extdata/ukcp18sst',package='mesoclim')
 sst<-create_ukcpsst_data(dir_data,as.POSIXlt('2018/05/01'),as.POSIXlt('2018/05/31'),members='01')
 dtmf<-rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
 dtmm<-rast(system.file('extdata/dtms/dtmm.tif',package='mesoclim'))
 climdata<-read_climdata(system.file('data/ukcpinput.rds',package='mesoclim'))
 tmf <- tempdownscale(climdata, sst, dtmf, dtmm, NA, NA,TRUE, TRUE,'tmax',2, 2)
 plot_q_layers(tmf)
}
