% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{tempelev}
\alias{tempelev}
\title{Downscale temperature with elevation effects}
\usage{
tempelev(tc, dtmf, dtmc = NA, rh = NA, pk = NA)
}
\arguments{
\item{tc}{a SpatRast or array of temperatures (deg C). If an array dtmc must
be provided.}

\item{dtmf}{a fine-resolution SpatRast of elevations. Temperatures down-scaled
to resolution of \code{dtmf}.}

\item{dtmc}{optional SpatRast of elevations matching resolution of \code{tc}. If
not supplied, \code{tc} must be a SpatRast and \code{dtmc} is derived by resampling \code{dtmf}
to resolution of \code{tc} and extents must match.}

\item{rh}{optional SpatRast or array of relative humidities (percentage). If
supplied, a humidity-dependent lapse rate is calculated. If not supplied, a
fixed lapse rate of -0.05 deg C/m is applied.}

\item{pk}{optional SpatRast or array of atmospheric pressures (kPa). Needed if
lapse-rate humidity dependent.}
}
\value{
a multi-layer SpatRast of elevation-corrected temperatures (deg C) matching
the resolution, extent and crs of dtmf.
}
\description{
Downscales an array of temperature data applying elevation effects
}
