% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdownscale.R
\name{temphrly_downscale}
\alias{temphrly_downscale}
\title{Downscale hourly temperature}
\usage{
temphrly_downscale(
  climhrly,
  sst,
  dtmf,
  dtmm = NA,
  basins = NA,
  uzf = NA,
  cad = TRUE,
  coastal = TRUE,
  thgto = 2,
  whgto = 2,
  tempvar = "temp"
)
}
\arguments{
\item{sst}{a SpatRast of sea-surface temperature data (deg C) timeseries that overlaps climdata$tme}

\item{dtmf}{a high-resolution SpatRast of elevations}

\item{dtmm}{a medium-resolution SpatRast of elevations covering a larger area
than dtmf (only needed for coastal effects - see details).}

\item{basins}{optionally, a fine-resolution SpatRast of basins as returned by \code{\link[=basindelin]{basindelin()}}
matching the coordinate reference system and extent of \code{dtmf}. Calculated if
not supplied.}

\item{uzf}{optionally, a SpatRast of fine resolution wind speeds as returned by \code{\link[=winddownscale]{winddownscale()}}.
Calculated if not supplied.}

\item{cad}{optional logical indicating whether to calculate cold-air drainage effects}

\item{coastal}{optional logical indicating whether to calculate coastal effects}

\item{thgto}{height above ground of output temperature measurements.}

\item{whgto}{height above ground of output wind speed measurements.}

\item{tempvar}{string name of element of climdata holding temperature data (default = 'temp')}

\item{climdata}{a \code{climdata} model object containing climate data of the same format as \code{era5climdata}.
Expects temperature to be in degrees and with the name \code{temp}.}
}
\value{
a multi-layer SpatRast of downscaled temperatures (deg C) matching the
resolution of dtmf.
}
\description{
Downscales coarse-resolution temperature data accounting for elevation
effects and optionally cold air drainage and coastal effects
}
\details{
Cold air drainage is calculated by delineating hydrological basins and
calculating flow accumulation and the elevation difference from the highest point of the basin.
Cold-air drainage is assumed to occur when atmospheric stability is high, namely when
the radiation balance is negative and wind speeds are low. Coastal effects are
calculated by determining the ratio of land to sea pixels in an upwind direction.
The provision of \code{dtmm} allows this ratio to be derived accoutning for land and
sea outside the boundaries of the study area.
}
\examples{
climdata<- read_climdata(mesoclim::ukcpinput)
climhrly<-temporaldownscale(climdaily, adjust = TRUE, clearsky=NA, srte = 0.09, relmin = 10, noraincut = 0)
dtmf<-terra::rast(system.file("extdata/dtms/dtmf.tif",package="mesoclim"))
dtmm<-terra::rast(system.file("extdata/dtms/dtmm.tif",package="mesoclim"))
basins<-basindelin(dtmf, boundary = 2)
wsfhr<-winddownscale(climhrly$windspeed, climhrly$winddir, dtmf, dtmm, climhrly$dtm, zi=climhrly$windheight_m)
sst<-terra::unwrap(ukcp18sst)
tempf<-temphrly_downscale(climhrly, sst, dtmf, dtmm, basins, uzf = wsfhr,tempvar="temp")
terra::panel(tempf[[c(1,6,12,18)]],main=paste('Temperature',c('00:00', '06:00', '12:00', '18:00'),' on 1 May 2018'))
matplot(terra::time(tempf),unlist(terra::global(tempf,mean,na.rm=TRUE)), type = "l", lty = 1, main='Mean hourly temperature of whole area')
}
\keyword{spatial}
