% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcp_functions.R
\name{ukcp18toclimarray}
\alias{ukcp18toclimarray}
\title{convert UKCP18  ncdf4 files to format required for model}
\usage{
ukcp18toclimarray(
  dir_ukcp,
  dtm,
  startdate,
  enddate,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  wsalbedo = 0.19,
  bsalbedo = 0.22,
  ukcp_vars = c("clt", "hurs", "huss", "pr", "psl", "rls", "rss", "tasmax", "tasmin",
    "uas", "vas"),
  ukcp_units = c("\%", "\%", "1", "mm/day", "hPa", "watt/m^2", "watt/m^2", "degC",
    "degC", "m/s", "m/s"),
  output_units = c("\%", "\%", "1", "mm/day", "kPa", "watt/m^2", "watt/m^2", "degC",
    "degC", "m/s", "m/s"),
  toArrays = TRUE,
  sampleplot = FALSE
)
}
\arguments{
\item{dir_ukcp}{directory holding ALL ncdf files of UKCP18 data required to extract the requested variables and timeseries.}

\item{dtm}{SpatRaster object of elevations covering the geographical extent required (see details)}

\item{startdate}{POSIXlt class defining start date of required timeseries}

\item{enddate}{POSIXlt class defining end date of required timeseries}

\item{collection}{text string defining UKCP18 collection, either 'land-gcm' or 'land-rcm'}

\item{domain}{text string defining UKCP18 domain, either 'uk' or 'eur'(land-rcm collection only) or 'global'}

\item{member}{string defining the climate model member to be used for the timeseries. Available members vary between UKCP18 collections.}

\item{ukcp_vars}{UKCP18 variable names to be extracted DO NOT CHANGE}

\item{ukcp_units}{units of the UKCP18 variables extracted DO NOT CHANGE}

\item{output_units}{units required for output DO NOT CHANGE CHECK CORRECT}

\item{toArrays}{logical determining if climate data returned as list of arrays. If FALSE returns list of Spatrasts.}

\item{sampleplot}{if TRUE plots examples of interpolated dates when converting from 360 to 366 day years}
}
\value{
a list of the following:
\describe{
\item{tme}{POSIXlt object of times corresponding to climate observations}
\item{climarray}{ either a list of arrays or a list of SpatRasts of daily weather variables
\item{dtmc}{a coarse resolution digital elevation dataset matching the resolution and crs of the input
climate data, and an extent matching \code{dtm}}
}
}
\description{
Converts UKCP18 global or regional land data in the form of netCDF4 file (as returned
by \code{\link[=ukcp18_downscale]{ukcp18_downscale()}} to the correct data form required for subsequent modelling.
}
\details{
The function converts 360 day UKCP18 to actual calendar data by reassignment and linear interpolation of missing data. Returned climate data will be at the resolution, corrdinate reference system of the UKCP18 data requested, which can vary between domains and collections.
The extent will be provided by the \code{dtm} which, if necessary, will be re-projected and resampled to the CRS and resolution of UKCP18 data requested.
For regional UKCP18 data, it is recommended that 'dtm' is derived from the original orography data available for download.
}
\examples{
dir_out <- tempdir()
download_ukcp18(dir_out,'ceda_username','ceda_password',as.POSIXlt('2018-05-01'),as.POSIXlt('2018-05-31'),'land-rcm','uk','rcp85',c('01','02','03'),c('clt','hurs','huss','pr','psl','rls','rss','tas','tasmax','tasmin','uas','vas'))
climdata<-ukcp18toclimarray()
}
