% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukcpfuture}
\alias{ukcpfuture}
\title{A list of 12km resolution UKCP18 model future climate data}
\format{
a list with the following elements:
\describe{
\item{dtm}{a wrapped SpatRast object of elevations (m) matching the extent and resolution of climate data}
\item{tme}{POSIXlt object of dates and times}
\item{windheight_m}{numeric value in metres of wind height above ground}
\item{tempheight_m}{numeric value in metres of temperature height above ground}
\item{tmax}{Maximum Daily Temperature (deg C)}
\item{tmin}{Minimum Daily Temperature (deg C)}
\item{relhum}{Relative humidity (Percentage)}
\item{spechum}{Specific humidity}
\item{pres}{Sea-level atmospheric pressure (kPa)}
\item{swnet}{Net downward shortwave radiation (W/m^2)}
\item{lwnet}{Net downward longwave radiation (W/m^2)}
\item{windspeed}{at 2m (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
\item{prec}{Precipitation (mm)}
}
}
\source{
\url{ftp.ceda.ac.uk}
}
\usage{
ukcpfuture
}
\description{
A list of hourly UKCP18 regional climata data for May 2030 for Southwest UK as
returned by \code{\link[=ukcp18toclimarray]{ukcp18toclimarray()}} where collection='land-gcm' & domain='uk'
}
\keyword{datasets}
